Compute sufficient statistics for input to susie_ss

Computes the sufficient statistics \(X'X, X'y, y'y\) and \(n\) after centering (and possibly standardizing) the columns of \(X\) and centering \(y\) to have mean zero. We also store the column means of \(X\) and mean of \(y\).

Usage

compute_suff_stat(X, y, standardize = FALSE)

Arguments

X

An n by p matrix of covariates.

y

An n vector.

standardize

Logical flag indicating whether to standardize columns of X to unit variance prior to computing summary data

Value

A list of sufficient statistics (XtX, Xty, yty, n) and X_colmeans, y_mean.

Examples

data(N2finemapping)
ss <- compute_suff_stat(N2finemapping$X, N2finemapping$Y[, 1])