Predict outcomes or extract coefficients from susie fit.

Predict outcomes or extract coefficients from susie fit.

# S3 method for susie
predict(object, newx = NULL, type = c("response", "coefficients"), ...)

Arguments

object

A susie fit.

newx

A new value for X at which to do predictions.

type

The type of output. For type = "response", predicted or fitted outcomes are returned; for type = "coefficients", the estimated coefficients are returned.

...

Other arguments used by generic predict function. These extra arguments are not used here.

Value

For type = "response", predicted or fitted outcomes are returned; for type = "coefficients", the estimated coefficients are returned. If the susie fit has intercept = NA (which is common when using susie_suff_stat) then predictions are computed using an intercept of 0, and a warning is emitted.