sets three attributes for matrix X

sets three attributes for matrix X

set_X_attributes(X, center = TRUE, scale = TRUE)

Arguments

X
an n by p data matrix that can be either a trend filtering matrix or a regular dense/sparse matrix

center
boolean indicating centered by column means or not

scale
boolean indicating scaled by column standard deviations or not

Value

X with three attributes e.g. attr(X, 'scaled:center') is a p vector of column means of X if center=TRUE, a p vector of 0s otherwise. attr(X, 'scaled:scale') is a p vector of column standard deviations of X if scale=TRUE, a p vector of 1s otherwise. attr(X, 'd') is a p vector of column sums of X.standardized^2, where X.standardized is the matrix X centered by attr(X, 'scaled:center') and scaled by attr(X, 'scaled:scale').