Bayesian single-effect linear regression using z scores.

Bayesian single-effect linear regression using z scores.

single_effect_regression_rss(
  z,
  Sigma,
  V = 1,
  prior_weights = NULL,
  optimize_V = c("none", "optim", "uniroot", "EM", "simple")
)

Arguments

z
a p vector

Sigma
residual_var * R + lambda I

V
the prior variance

prior_weights
a p vector of prior weights

optimize_V
boolean indicating whether to optimize V (by maximum likelihood)

Value

a list with elements:

alpha

vector of posterior inclusion probabilities. ie alpha[i] is posterior probability that that b[i] is non-zero

mu

vector of posterior means (conditional on inclusion)

mu2

vector of posterior second moments (conditional on inclusion)

lbf

vector of log Bayes factors for each variable

V

the prior variance (after optimization, if optimize_V is TRUE)

lbf_model

(scalar) the loglikelihood for the total model minus the log-likelihood for the null model

Details

Performs single-effect linear regression with z scores. That is, this function fits the regression model z = Rb + e, where e is N(0,Sigma), Sigma = residual_var * R + lambda I, and the b is a p vector of effects to be estimated. The assumption is that b has exactly one non-zero element, with all elements equally likely to be non-zero. The prior on the non-zero element is N(0,var=V).

Performs sum of single-effect (susie) linear regression with z scores. The summary data required are the p by p correlation matrix R, the p vector z. The summary stats should come from the same individuals. This function fits the regression model z = sum_l Rb_l + e, where e is N(0,Sigma) and the sum_l b_l is a p vector of effects to be estimated. The assumption is that each b_l has exactly one non-zero element, with all elements equally likely to be non-zero. The prior on the non-zero element is N(0,var=prior_variance).