An attempt to automate reliable running of susie even on hard problems

An attempt to automate reliable running of susie even on hard problems

susie_auto(X, Y, L_init = 1, L_max = 512, verbose = FALSE,
  init_tol = 1, standardize = TRUE, intercept = TRUE,
  max_iter = 100, tol = 0.01, ...)

Arguments

X
an n by p matrix of covariates

Y
an n vector

L_init
the initial value of L to consider

L_max
the maximum value of L to consider

verbose
if true outputs some progress messages

init_tol
the tolerance to pass to susie during early runs (set big to run faster)

standardize
logical flag for whether to standardize columns of X to unit variance prior to fitting. Note that `prior_variance` specifies the prior on the coefficients of X after standardization (if performed). If you do not standardize you may need to think carefully about specifying `prior_variance`. Whatever the value of standardize, the coefficients (returned from `coef`) are for X on the original input scale. Any column of X that has zero variance is not standardized, but left as is.

intercept
Should intercept be fitted (default=TRUE) or set to zero (FALSE)

max_iter
maximum number of iterations to perform

tol
convergence tolerance

Details

Currently runs a 3-stage strategy for each L: first fit susie with very small residual error, then estimate residual error, then estimate prior variance. If the last step estimates some prior variances to be 0 then stop. Otherwise double L and repeat. Initial runs are done with lax tolerance (init_tol); final run done with default tolerance.