Extract credible sets from SuSiE fit

Extract credible sets from SuSiE fit

susie_get_cs(res, X = NULL, Xcorr = NULL, coverage = 0.95,
  min_abs_corr = 0.5, dedup = TRUE)

Arguments

res
a susie fit, the output of `susieR::susie()`, or simply the posterior inclusion probability matrix alpha.

X
N by P matrix of values of the P variables (covariates) in N samples. When provided, correlation between variables will be computed and used to remove CSs whose minimum correlation among variables is smaller than `min_abs_corr` (see below).

Xcorr
P by P matrix of correlations between variables (covariates). When provided, it will be used to remove CSs whose minimum correlation among variables is smaller than `min_abs_corr` (see below).

coverage
a number (in [0,1]) specifying desired coverage of each CS

min_abs_corr
a "purity" threshold for the CS. Any CS that contains a pair of variables with correlation less than this threshold will be filtered out and not reported.

dedup
If TRUE, "deduplicate" - that is remove duplicated CSs.

Value

a list with elements:

cs

a list, each element corresponds to a CS, and is a vector containing the indices of the variables in the CS.

coverage

the nominal coverage specified for each CS.

purity

(If `X` or `Xcorr` are provided), the purity of each CS.

cs_index

(If `X` or `Xcorr` are provided) the index (in 1,...,L) of each reported CS in the supplied susie fit.

Details

Reports indices of variables in each credible set (CS) identified, as well as summaries of correlation among the variables included in each CS. If desired, one can filter out CSs that do not meet a specified purity threshold (min_abs_corr); to do this either `X` or `Xcorr` must be supplied.