Plot per variable summary in SuSiE CSs

Plot per variable summary in SuSiE CSs

susie_plot(model, y, add_bar = FALSE, pos = NULL, b = NULL,
  max_cs = 400, add_legend = FALSE, ...)

Arguments

model
a susie fit, the output of `susieR::susie()`. It has to contain `z`, `PIP` and optionally `sets`. It is also possible to take in a vector of z-score or PIP, in order to plot data from other software program.

y
a string indicating what to plot: z (for z-score), PIP, log10PIP or a random label to plot input data as is.

add_bar
add horizontal bar to signals in credible interval.

pos
coordinates of variables to plot, default to all variables

b
for simulated data, specify b = true effects (highlights in red).

max_cs
the biggest CS to display, based on purity (set max_cs in between 0 and 1) or size (>1).

add_legend
if TRUE, add a legend to annotate the size and purity of each CS discovered.