Plot changepoint data and susie fit using ggplot2

Plot changepoint data and susie fit using ggplot2

susie_plot_changepoint(s, y, line_col = "blue", line_size = 1.5,
  cs_col = "red")

Arguments

s
a susie fit obtained by applying susie_trendfilter(y,order=0) to y

y
an n vector of observations that are ordered in time or space (assumed equally-spaced)

line_col
color for the line showing fitted values

line_size
a size for the line

cs_col
color for the shaded rectangles showing credible sets

Value

a ggplot2 object for plotting

Details

Plots original data, y, overlaid with line showing susie fitted value and shaded rectangles showing credible sets for changepoint locations

Examples

set.seed(1)
mu = c(rep(0,50),rep(1,50),rep(3,50),rep(-2,50),rep(0,300))
y = mu + rnorm(500)
s = susie_trendfilter(y)
susie_plot_changepoint(s,y) # produces ggplot with credible sets for changepoints on top of plot