Bayesian sum of single-effect (susie) linear regression using summary stat

Bayesian sum of single-effect (susie) linear regression using summary stat

susie_ss(XtX, Xty, var_y = 1, n, L = 10, scaled_prior_variance = 0.2,
  residual_variance = NULL, estimate_prior_variance = FALSE,
  max_iter = 100, s_init = NULL, verbose = FALSE, intercept = 0,
  tol = 0.0001)

Arguments

XtX
a p by p matrix, X'X, where columns of X are centered to have mean 0

Xty
a p vector, X'Y, where columns of X are centered and Y is centered to have mean 0

var_y
the (sample) variance of the vector Y

n
sample size

L
maximum number of non-zero effects

scaled_prior_variance
the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times )

residual_variance
the residual variance (defaults to variance of Y)

estimate_prior_variance
indicates whether to estimate prior (currently not recommended as not working as well)

max_iter
maximum number of iterations to perform

s_init
a previous susie fit with which to initialize

intercept
a value to assign to the intercept (since the intercept cannot be estimated from centered summary data). This value will be used by coef.susie() to assign an intercept value, for consistency with the non-summary-statistic version of this function susie. Set to NULL if you want coef.susie() not to include an intercept term (and so only return a p vector).

tol
convergence tolerance based on alpha

Value

a susie fit, which is a list with some or all of the following elements

alpha

an L by p matrix of posterior inclusion probabilites

mu

an L by p matrix of posterior means (conditional on inclusion)

mu2

an L by p matrix of posterior second moments (conditional on inclusion)

XtXr

an p vector of t(X) times fitted values, the fitted values equal to X times colSums(alpha*mu))

sigma2

residual variance

V

prior variance

Details

Performs sum of single-effect (susie) linear regression of Y on X when only summary statistics are available. The summary data required are the p by p matrix X'X, the p vector X'Y, and the sample variance of Y, or (1/n)Y'Y. The summary stats should come from the same individuals. Both the columns of X and the vector Y should be centered to have mean 0 before computing these summary statistics; you may also want to scale each column of X and Y to have variance 1 (see examples). This function fits the regression model Y = sum_l Xb_l + e, where elements of e are iid N(0,var=residual_variance) and the sum_l b_l is a p vector of effects to be estimated. The assumption is that each b_l has exactly one non-zero element, with all elements equally likely to be non-zero. The prior on the non-zero element is N(0,var=scaled_prior_variance*var(Y)).

Examples

set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1:4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = c(X %*% beta + rnorm(n))
X = scale(X,center=TRUE, scale=TRUE)
y = (y - mean(y))/sd(y)
res =susie_ss(XtX = crossprod(X),Xty= c(y %*% X), var_y=1, n=1000)
coef(res)

#>    [1]  0.000e+00  4.565e-01  4.420e-01  4.302e-01  4.268e-01 -4.728e-05
#>    [7] -4.576e-05  2.169e-04  1.812e-05  5.685e-05  1.197e-05  8.827e-05
#>   [13]  1.268e-04  7.703e-05  1.533e-04  1.857e-05 -4.374e-05  2.218e-04
#>   [19]  7.865e-05  8.422e-05  2.391e-06 -3.592e-05 -7.973e-05  1.294e-04
#>   [25]  3.127e-05  1.556e-04  1.087e-05 -5.719e-05 -5.560e-05  6.937e-05
#>   [31] -1.169e-04 -2.214e-05  1.351e-04  1.383e-04  1.402e-05  2.436e-05
#>   [37]  1.237e-04 -3.422e-05 -7.164e-05  3.921e-05  1.190e-04  5.916e-05
#>   [43]  2.110e-05 -2.287e-05 -4.567e-05 -1.094e-04  1.415e-04 -2.140e-05
#>   [49]  1.899e-04  1.966e-05 -6.350e-05 -3.110e-05  9.674e-05 -1.470e-04
#>   [55]  2.060e-05  8.475e-05  4.381e-05  9.864e-05 -4.403e-05 -2.486e-05
#>   [61]  4.187e-04  4.374e-05  3.174e-05 -2.269e-05 -6.370e-05  5.128e-06
#>   [67] -7.853e-05  4.467e-05 -4.462e-05 -4.566e-04 -4.398e-06  1.426e-04
#>   [73]  9.711e-05  2.593e-05  7.480e-05 -9.934e-06 -4.969e-05  2.657e-05
#>   [79]  1.491e-04  1.091e-04 -1.740e-04  5.533e-05 -1.718e-06 -7.801e-06
#>   [85] -3.980e-05  9.008e-05  7.653e-05  2.435e-05 -1.690e-05 -1.877e-05
#>   [91]  9.621e-05 -1.346e-04  2.506e-05  2.797e-05  3.070e-05 -1.210e-04
#>   [97]  1.159e-04 -6.108e-05  1.082e-04 -1.655e-04 -1.822e-05  1.053e-04
#>  [103]  1.087e-04 -5.526e-05  6.770e-06  1.278e-04  6.451e-05 -2.516e-05
#>  [109]  7.900e-06  3.779e-05  8.548e-05  9.603e-05  5.833e-05 -8.246e-05
#>  [115]  2.402e-05  2.348e-05  2.350e-05 -1.638e-05 -1.147e-05  1.346e-05
#>  [121]  1.697e-04  7.476e-05  4.068e-05  3.338e-05 -5.030e-05  4.182e-05
#>  [127] -4.743e-05 -1.010e-04  2.474e-04  2.354e-05  7.502e-05  1.062e-04
#>  [133] -6.781e-06 -2.811e-05 -4.225e-05  1.640e-05  2.376e-05  1.461e-05
#>  [139] -4.069e-05 -4.781e-05 -4.700e-05  7.196e-05 -2.083e-05 -8.127e-05
#>  [145]  8.482e-07  5.819e-06 -7.215e-05  1.526e-04  1.584e-05  1.687e-04
#>  [151] -3.338e-05 -4.347e-05  5.303e-05  1.281e-04 -5.654e-05 -2.062e-05
#>  [157] -2.302e-04 -5.227e-05  5.244e-05  2.355e-05 -8.352e-05  9.216e-06
#>  [163] -2.055e-05 -2.159e-05  1.341e-04 -6.146e-05  9.380e-05  1.208e-04
#>  [169]  9.845e-05  6.664e-05  2.587e-04  2.071e-05  1.014e-04  1.052e-04
#>  [175] -1.919e-06 -5.074e-04  1.667e-05 -2.266e-06  9.695e-05 -2.190e-05
#>  [181]  1.171e-04 -3.670e-05 -3.561e-05  5.829e-06  8.437e-05  1.265e-04
#>  [187] -5.596e-05 -3.012e-05 -4.100e-05 -1.064e-04 -9.735e-05 -4.915e-05
#>  [193]  1.185e-05  1.267e-04  6.524e-05 -6.379e-06 -2.395e-05 -8.750e-05
#>  [199] -4.861e-05  2.591e-05 -1.797e-05  8.194e-06 -1.235e-04  2.255e-05
#>  [205] -2.038e-04  4.376e-05  5.370e-06  5.200e-05  5.959e-05  1.426e-05
#>  [211] -4.353e-05  2.853e-06  6.818e-05 -1.304e-04 -7.671e-05 -4.982e-05
#>  [217]  5.052e-05 -4.215e-05 -1.068e-04  2.213e-04  5.277e-05 -3.431e-05
#>  [223]  4.869e-05 -6.412e-05  7.065e-05  5.617e-06 -3.069e-05  7.256e-05
#>  [229] -4.600e-05  5.791e-05 -3.492e-05  6.458e-05 -2.091e-04  4.993e-05
#>  [235]  2.245e-05  3.450e-05  2.312e-05 -6.595e-05  2.828e-05  1.595e-06
#>  [241]  4.715e-05  2.731e-05  1.516e-05  7.337e-05  6.608e-05  7.668e-06
#>  [247]  1.244e-04 -1.916e-05 -8.516e-05 -2.959e-05 -6.323e-05  9.768e-06
#>  [253]  6.331e-05  4.433e-05 -1.969e-05 -1.007e-04  1.874e-04  6.858e-07
#>  [259]  9.595e-05  4.082e-06  5.619e-05 -3.992e-05 -3.305e-06  4.416e-05
#>  [265] -1.847e-04  6.299e-05 -1.924e-05 -5.511e-05  7.233e-05  1.395e-04
#>  [271]  2.071e-04  2.340e-05 -1.007e-04 -1.483e-04 -3.439e-06  3.410e-04
#>  [277] -2.618e-05  2.711e-05 -1.154e-05 -1.034e-04  6.414e-05  2.018e-05
#>  [283] -2.426e-05 -1.108e-04 -1.420e-04  3.371e-06  5.191e-05 -1.400e-05
#>  [289]  3.515e-05  1.786e-07 -2.008e-05 -6.861e-05  1.535e-04  1.986e-05
#>  [295]  6.543e-06 -5.801e-05 -1.454e-04 -6.635e-05 -6.958e-05 -6.324e-06
#>  [301] -3.571e-05 -3.348e-05 -2.498e-05  3.121e-04 -1.441e-04  4.405e-05
#>  [307] -1.951e-05 -6.000e-06 -1.396e-04 -8.739e-05  8.071e-05  7.137e-05
#>  [313]  2.540e-06 -6.298e-06 -9.289e-06  1.224e-04  1.402e-04  1.814e-05
#>  [319] -1.640e-05  3.227e-04 -7.871e-05 -4.520e-05 -4.390e-05  1.223e-06
#>  [325]  3.211e-05  9.662e-07  1.300e-04 -5.450e-05  1.568e-05  4.720e-05
#>  [331] -4.377e-05 -1.424e-06  7.702e-05  1.094e-04 -1.869e-05 -5.213e-05
#>  [337]  1.183e-04 -1.192e-04  7.303e-05  5.200e-05 -1.580e-05  5.681e-05
#>  [343]  8.739e-06 -1.122e-05 -1.936e-04  1.927e-05  3.414e-04 -4.117e-05
#>  [349] -2.500e-05 -2.264e-05  7.572e-05  7.390e-05 -3.028e-05  3.243e-05
#>  [355] -6.735e-05 -4.182e-05 -6.877e-05 -3.790e-05 -2.595e-05 -1.251e-04
#>  [361] -1.622e-05 -6.485e-05  7.094e-05 -5.532e-05 -1.328e-04 -1.241e-04
#>  [367] -4.837e-05 -1.426e-05  1.152e-04  3.360e-05  4.048e-05 -3.826e-05
#>  [373] -1.208e-04  4.071e-05 -2.103e-04  3.107e-05  4.134e-05 -1.495e-04
#>  [379]  3.925e-05 -9.516e-05  9.061e-05  9.274e-06  2.178e-05  1.581e-04
#>  [385]  5.491e-05 -9.051e-05 -3.484e-05 -5.679e-05 -1.764e-04  2.647e-05
#>  [391]  4.709e-05  1.263e-04  4.974e-05 -1.331e-05 -1.502e-05  1.370e-05
#>  [397] -3.349e-05 -1.625e-04  2.494e-05 -6.330e-05  3.235e-05 -1.322e-05
#>  [403] -3.535e-05 -2.376e-05 -1.205e-05  8.234e-05 -6.496e-05  6.270e-05
#>  [409]  5.858e-05 -1.641e-04 -2.419e-05 -5.918e-05  4.643e-05  5.545e-05
#>  [415]  9.111e-06 -8.745e-05 -5.599e-05 -7.549e-05  4.981e-05  2.455e-05
#>  [421] -1.852e-04  1.267e-04 -1.533e-05  3.311e-05  7.217e-07 -8.276e-05
#>  [427] -1.515e-04  5.657e-05 -6.991e-05  7.738e-05 -5.194e-05 -1.267e-05
#>  [433] -3.564e-05  1.141e-04 -6.017e-05  1.486e-04  2.568e-05  9.977e-05
#>  [439]  3.302e-06 -1.746e-05  6.301e-05  2.536e-05  5.852e-05  1.396e-05
#>  [445]  3.256e-05 -1.304e-04 -5.546e-05  1.043e-05 -6.998e-05  2.907e-06
#>  [451]  2.234e-05 -1.969e-04 -4.714e-05  8.114e-05 -3.953e-05  5.690e-05
#>  [457]  2.845e-05  4.016e-05 -6.172e-05  8.363e-05  2.917e-05  1.289e-04
#>  [463]  1.073e-04  2.742e-05  3.168e-05  7.104e-05  4.399e-05  3.132e-05
#>  [469]  7.711e-05  2.765e-04  3.024e-04 -7.453e-05 -7.833e-06 -1.848e-05
#>  [475] -3.934e-05 -1.440e-05 -6.493e-05 -5.969e-05  7.741e-05  2.389e-05
#>  [481]  3.008e-05  1.418e-05 -1.589e-04  3.602e-05 -1.498e-04  1.566e-05
#>  [487] -3.406e-04 -7.610e-05 -1.124e-04 -1.000e-04 -1.701e-04  2.575e-05
#>  [493] -7.849e-05 -5.398e-05 -2.218e-05  1.895e-06  5.111e-05 -4.751e-05
#>  [499]  2.105e-05 -7.045e-05  3.124e-05 -1.514e-05 -4.881e-05 -2.003e-05
#>  [505] -7.497e-05 -4.749e-06 -1.999e-05 -5.767e-05  1.115e-04 -3.171e-05
#>  [511]  6.331e-08 -1.005e-04 -6.662e-05  5.886e-05 -1.704e-04 -7.705e-05
#>  [517] -7.570e-05  5.905e-05  1.163e-05 -3.450e-05 -1.092e-05  1.255e-04
#>  [523] -2.051e-04  3.247e-05  5.882e-05 -6.863e-05  8.805e-05 -7.343e-06
#>  [529]  3.398e-06  1.003e-04 -2.675e-05  1.422e-04  1.177e-04 -2.229e-05
#>  [535] -1.077e-05  6.110e-06  1.903e-05 -5.003e-05  4.450e-05 -2.031e-04
#>  [541] -4.517e-05  5.176e-05  2.374e-05 -7.840e-05 -9.895e-05  7.822e-05
#>  [547] -4.931e-05 -8.621e-05  1.471e-05  1.005e-04  2.152e-05  1.692e-05
#>  [553] -1.383e-05 -3.777e-06 -1.366e-04 -7.053e-04  1.516e-04 -8.976e-05
#>  [559]  1.033e-04  5.759e-05  1.479e-04  1.013e-04 -6.783e-05  9.190e-06
#>  [565]  9.259e-05  5.869e-05 -6.965e-05 -5.461e-05 -2.272e-04  3.296e-05
#>  [571]  6.530e-05 -8.796e-05  5.904e-06 -1.860e-04  1.397e-05  1.042e-04
#>  [577] -5.307e-05  1.735e-05  1.353e-05  6.446e-05  6.648e-05  4.161e-05
#>  [583]  9.618e-05  4.464e-05  2.935e-05  1.439e-05 -1.118e-04  3.736e-05
#>  [589]  6.199e-05 -3.436e-05  9.538e-05  1.124e-04 -1.176e-04  1.440e-04
#>  [595]  1.270e-04 -5.681e-05 -1.381e-04 -4.607e-06 -5.316e-05 -1.701e-05
#>  [601]  1.283e-04  9.393e-05  3.172e-05  5.509e-05  5.046e-05  8.358e-05
#>  [607] -8.020e-06 -3.076e-06  1.193e-05 -8.063e-05  6.326e-05  3.613e-05
#>  [613]  4.175e-04 -3.504e-05 -1.366e-04  4.496e-05 -5.947e-05 -2.256e-05
#>  [619]  9.445e-05 -6.718e-06  3.592e-05  1.361e-04  6.637e-05 -6.336e-07
#>  [625] -1.079e-04 -2.602e-05 -6.569e-05  1.183e-04  1.419e-04 -9.621e-05
#>  [631] -7.428e-05  2.269e-05 -3.003e-05  1.010e-05 -8.834e-06  4.135e-05
#>  [637] -4.074e-04  6.491e-05 -8.275e-05  7.960e-05  2.019e-05  2.989e-04
#>  [643] -3.941e-05 -1.557e-06 -9.565e-05  1.680e-05 -9.339e-05  1.084e-04
#>  [649] -7.760e-06  4.455e-06 -8.694e-05  3.294e-05 -1.464e-04 -7.731e-06
#>  [655] -8.365e-05  2.815e-05 -1.060e-04  1.518e-05  3.908e-05 -5.439e-05
#>  [661]  1.516e-05 -5.151e-05  5.628e-06 -4.115e-05  4.323e-05  5.461e-06
#>  [667] -3.864e-05  4.045e-06  3.897e-05 -1.117e-06  2.318e-05  1.016e-04
#>  [673] -5.406e-05  4.961e-05 -5.025e-05  1.644e-04 -6.006e-05  1.036e-05
#>  [679] -4.148e-05 -7.357e-06  2.027e-04  3.264e-05  1.777e-04 -1.235e-04
#>  [685] -9.997e-05 -1.898e-05 -3.512e-05 -3.289e-05 -1.699e-05  1.718e-04
#>  [691]  3.487e-05  1.365e-04  3.686e-05 -6.619e-05  2.897e-05  7.856e-06
#>  [697] -2.698e-05  1.032e-04 -5.693e-05  9.096e-06 -1.352e-05  3.302e-06
#>  [703] -8.683e-05 -9.019e-05  4.740e-05 -3.628e-06  2.649e-05  1.264e-04
#>  [709]  4.757e-05 -8.146e-05 -5.665e-05  3.440e-05  4.786e-04  4.365e-05
#>  [715] -1.151e-04  9.442e-06 -8.991e-05  1.093e-04  1.976e-05 -6.619e-06
#>  [721] -9.947e-05  3.295e-05  1.271e-04  1.078e-05  4.267e-05  3.925e-06
#>  [727]  1.546e-05 -3.918e-04 -2.819e-05 -2.411e-05  2.936e-05 -3.139e-05
#>  [733]  2.347e-04 -1.081e-05 -6.735e-05 -8.741e-05 -1.440e-05 -6.870e-05
#>  [739]  7.631e-05 -1.583e-04 -2.171e-05 -2.275e-05  2.506e-05  3.496e-05
#>  [745] -4.824e-05  4.424e-05  3.303e-07 -6.155e-06  3.206e-05 -3.783e-05
#>  [751] -4.448e-05  7.138e-05 -1.489e-04  2.869e-05 -2.220e-05 -3.783e-06
#>  [757] -2.708e-05 -1.497e-04  5.556e-05 -2.762e-06  6.393e-05  4.533e-05
#>  [763] -1.505e-05 -2.378e-04  1.198e-04 -1.637e-04 -3.661e-05  4.394e-05
#>  [769]  6.112e-05  2.603e-05  2.745e-04  3.299e-06  2.046e-04 -7.012e-05
#>  [775]  2.587e-04 -1.759e-05 -1.571e-05  6.178e-05 -1.375e-04 -7.745e-05
#>  [781]  4.315e-06  2.521e-05  2.934e-05  1.123e-04  3.715e-05 -8.991e-05
#>  [787] -6.121e-05  2.351e-05 -2.357e-05 -1.564e-04  8.815e-05 -3.307e-05
#>  [793]  9.736e-05 -2.804e-05 -1.277e-05  1.138e-04 -4.811e-05 -1.933e-04
#>  [799] -9.301e-06  4.616e-05 -7.344e-05  3.320e-05  7.451e-05  5.632e-05
#>  [805]  3.018e-05 -1.307e-04 -3.041e-05  5.760e-05  4.291e-05 -3.125e-05
#>  [811] -2.990e-05  3.757e-05  1.465e-05  3.547e-05  1.632e-05 -5.233e-05
#>  [817]  4.096e-05 -3.331e-05  1.275e-04  1.833e-05 -1.099e-04 -6.025e-05
#>  [823]  1.197e-05 -7.598e-06 -6.846e-05 -4.882e-05 -5.533e-06 -1.222e-04
#>  [829]  5.766e-05  4.627e-05  3.471e-05 -6.001e-05 -6.816e-05  9.864e-05
#>  [835] -7.158e-05  6.460e-05  1.855e-05  1.947e-05 -4.253e-05  5.316e-05
#>  [841] -3.413e-05 -5.887e-05 -6.493e-05  3.237e-05  5.956e-05  8.459e-05
#>  [847]  5.042e-05  1.821e-05 -2.682e-05 -5.387e-05 -1.549e-06 -2.663e-05
#>  [853]  6.406e-05 -8.968e-05  8.872e-05 -2.946e-05 -9.207e-05  2.119e-05
#>  [859]  2.113e-05  3.521e-05 -5.866e-05 -1.045e-04  1.828e-05  3.525e-05
#>  [865]  8.106e-07 -2.826e-05  1.196e-05 -3.487e-05 -1.919e-05 -2.396e-05
#>  [871]  8.042e-05 -1.815e-05 -2.468e-05 -3.492e-05 -4.503e-05  8.184e-05
#>  [877]  4.261e-05  4.766e-05  5.133e-05  1.940e-05  1.324e-05  2.225e-04
#>  [883] -9.023e-05 -2.077e-06  7.807e-05 -9.272e-05 -2.348e-05 -6.039e-05
#>  [889] -4.530e-05  2.477e-05  1.250e-05 -5.276e-06  8.825e-05 -1.392e-06
#>  [895] -1.812e-05 -1.205e-04  7.985e-05  1.471e-04 -2.485e-04 -2.747e-05
#>  [901] -3.195e-05  1.686e-04  7.020e-05  2.031e-04 -1.793e-05 -6.898e-05
#>  [907] -1.278e-04 -1.919e-05  2.160e-04 -5.053e-05  3.502e-05  4.571e-06
#>  [913] -1.671e-04 -1.387e-04  3.510e-05  9.077e-05  1.321e-05 -2.904e-05
#>  [919] -1.270e-04 -2.456e-06  4.690e-05  2.409e-05  5.560e-05  7.097e-05
#>  [925]  4.274e-05 -3.360e-05 -1.687e-05 -2.502e-05  1.474e-05 -9.253e-06
#>  [931] -1.644e-04  1.133e-04 -3.072e-05  1.142e-04 -3.959e-04  4.590e-05
#>  [937] -5.055e-05  2.902e-05 -1.041e-04  4.812e-05 -5.284e-05  4.453e-05
#>  [943]  8.300e-05  2.294e-05  7.079e-05 -7.786e-05 -4.148e-05  4.510e-06
#>  [949]  1.307e-04  1.648e-05  1.601e-04 -7.986e-06 -8.621e-05 -5.493e-08
#>  [955] -6.264e-05 -4.859e-05  1.884e-05 -8.938e-05 -1.145e-05 -1.218e-04
#>  [961] -4.137e-05  6.308e-05  6.244e-06  9.070e-05 -4.840e-05 -2.150e-04
#>  [967]  5.435e-05 -1.313e-04 -2.364e-05  9.970e-05  5.130e-06  4.184e-05
#>  [973]  4.190e-05 -9.047e-06 -6.562e-05 -2.346e-05  3.680e-06 -1.499e-04
#>  [979] -6.150e-05 -2.828e-05  1.222e-04 -5.705e-05 -9.880e-05 -5.935e-05
#>  [985] -3.608e-05 -6.036e-05  3.061e-05 -1.091e-04  3.459e-05 -5.843e-05
#>  [991]  5.878e-05  2.859e-05  1.485e-04 -5.806e-05 -5.134e-05 -2.185e-05
#>  [997]  1.903e-04 -2.894e-05  1.747e-04 -5.271e-05  1.475e-04