Bayesian sum of single-effect (susie) linear regression using sufficient summary stat

Bayesian sum of single-effect (susie) linear regression using sufficient summary stat

susie_suff_stat(
  bhat,
  shat,
  R,
  n,
  var_y = 1,
  XtX,
  Xty,
  yty,
  maf = NULL,
  maf_thresh = 0,
  L = 10,
  scaled_prior_variance = 0.2,
  residual_variance = NULL,
  estimate_residual_variance = TRUE,
  estimate_prior_variance = TRUE,
  estimate_prior_method = c("optim", "EM", "simple"),
  r_tol = 1e-08,
  prior_weights = NULL,
  null_weight = NULL,
  standardize = TRUE,
  max_iter = 100,
  s_init = NULL,
  intercept_value = 0,
  coverage = 0.95,
  min_abs_corr = 0.5,
  tol = 0.001,
  verbose = FALSE,
  track_fit = FALSE,
  check_input = FALSE
)

Arguments

bhat
a p vector of estimated effects.

shat
a p vector of corresponding standard errors.

R
a p by p symmetric and positive semidefinite matrix. It can be X'X, covariance matrix (X'X/(n-1)) or correlation matrix. It should from the same samples used to compute `bhat` and `shat`. Using out of sample matrix may produce unreliable results.

n
sample size.

var_y
the (sample) variance of y, defined as y'y/(n-1) . If it is unknown, the coefficients (returned from `coef`) are on the standardized X, y scale.

XtX
a p by p matrix, X'X, where columns of X are centered to have mean 0

Xty
a p vector, X'y, where columns of X are centered and y is centered to have mean 0

yty
a scaler, y'y, where y is centered to have mean 0

maf
minor allele frequency; to be used along with `maf_thresh` to filter input summary statistics

maf_thresh
variants having MAF smaller than this threshold will be filtered out

L
maximum number of non-zero effects

scaled_prior_variance
the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times)

residual_variance
the residual variance (defaults to variance of y)

estimate_residual_variance
indicates whether to estimate residual variance

estimate_prior_variance
indicates whether to estimate prior

estimate_prior_method
The method used for estimating prior variance.

r_tol
tolerance level for eigen value check of positive semidefinite matrix of R.

prior_weights
a p vector of prior probability that each element is non-zero

null_weight
probability of no effect, for each single effect model

standardize
logical flag (default=TRUE) for whether to adjust XtX and Xty such that they are computed from column standardized X, prior to fitting. Note that `scaled_prior_variance` specifies the prior on the coefficients of X *after* standardization (if performed). If you do not standardize you may need to think more carefully about specifying `scaled_prior_variance`. Whatever the value of standardize, the coefficients (returned from `coef`) are for X on the original input scale.

max_iter
maximum number of iterations to perform

s_init
a previous susie fit with which to initialize

intercept_value
a value to assign to the intercept (since the intercept cannot be estimated from centered summary data). This value will be used by coef.susie() to assign an intercept value, for consistency with the non-summary-statistic version of this function susie. Set to NULL if you want coef.susie() not to include an intercept term (and so only return a p vector).

coverage
coverage of confident sets. Default to 0.95 for 95% credible interval.

min_abs_corr
minimum of absolute value of correlation allowed in a credible set.

tol
convergence tolerance based on alpha

verbose
if true outputs some progress messages

track_fit
add an attribute trace to output that saves current values of all iterations

check_input
whether to perform checks on XtX and Xty, the checks are:

1. Check whether XtX is positive semidefinite

2. Check whether Xty in space spanned by the non-zero eigenvectors of XtX

Value

a susie fit, which is a list with some or all of the following elements

alpha

an L by p matrix of posterior inclusion probabilites

mu

an L by p matrix of posterior means (conditional on inclusion)

mu2

an L by p matrix of posterior second moments (conditional on inclusion)

XtXr

an p vector of t(X) times fitted values, the fitted values equal to X times colSums(alpha*mu))

sigma2

residual variance

V

prior variance

Details

Performs sum of single-effect (susie) linear regression of y on X when only sufficient summary statistics are available. The sufficient summary data required are EITHER the p vector bhat, the p vector shat, the p by p symmetric and positive semidefinite correlation (or covariance) matrix R, the sample size n, the variance of y; OR the p by p matrix X'X, the p vector X'y, the sum of squares of y (y'y) and the sample size. The sufficient summary stats should come from the same individuals. Both the columns of X and the vector y should be centered to have mean 0 before computing these summary statistics; you may also want to scale each column of X and y to have variance 1 (see examples). This function fits the regression model y = sum_l Xb_l + e, where elements of e are iid N(0,var=residual_variance) and the sum_l b_l is a p vector of effects to be estimated. The assumption is that each b_l has exactly one non-zero element, with all elements equally likely to be non-zero. The prior on the non-zero element is N(0,var=scaled_prior_variance*y'y/(n-1)).

Examples

set.seed(1)
n    <- 1000
p    <- 1000
beta <- rep(0,p)
beta[1:4] <- 1
X        <- matrix(rnorm(n*p),nrow=n,ncol=p)
y        <- c(X %*% beta + rnorm(n))
input_ss <- compute_ss(X,y,standardize = TRUE)
ss = susieR:::univariate_regression(X, y)
R = with(input_ss, cov2cor(XtX))
res1      <- with(input_ss,susie_suff_stat(XtX = XtX,Xty = Xty, yty = yty,n=n))
coef(res1)

#>    [1]   0.000000e+00   1.098768e+00   1.064003e+00   1.035326e+00
#>    [5]   1.027341e+00 -7.701477e-234 -1.724729e-230  1.036565e-231
#>    [9]  4.040300e-233  1.415289e-233 -4.010176e-234  1.724280e-233
#>   [13]  4.678108e-233  9.640184e-234  2.212882e-233 -1.007776e-233
#>   [17]  3.557835e-235  9.570969e-233  2.691977e-232  2.333910e-233
#>   [21] -3.044517e-233 -4.717275e-232 -1.522567e-232  2.212105e-232
#>   [25] -3.904739e-233  4.228649e-233 -1.216122e-233  8.699736e-234
#>   [29]  1.607525e-234  4.573805e-234 -1.041946e-233 -1.943015e-234
#>   [33]  9.012909e-231  6.400573e-234 -5.450719e-234 -3.810187e-234
#>   [37]  1.417560e-229  3.231910e-234 -3.270418e-233 -1.386600e-233
#>   [41]  3.880373e-233  4.296040e-234  6.786169e-235 -7.599220e-234
#>   [45] -8.121999e-233 -4.332351e-233  1.701318e-230 -6.640525e-232
#>   [49]  1.810728e-233 -2.164247e-234  3.533725e-232 -1.317365e-232
#>   [53] -2.451692e-234  5.406645e-235 -8.896427e-234  1.195703e-233
#>   [57]  4.270677e-233  5.651903e-234 -4.708835e-234 -5.676242e-233
#>   [61]  1.145972e-232  5.726108e-234 -2.978268e-232  3.143082e-234
#>   [65] -5.253710e-234 -2.460313e-234  4.735053e-234  1.087745e-232
#>   [69]  1.727780e-233 -1.234668e-229  4.526998e-234  7.038829e-233
#>   [73]  1.776170e-231  2.248523e-234 -6.173419e-232 -3.202678e-233
#>   [77] -3.353558e-234  9.484692e-234  1.333334e-232  4.238471e-231
#>   [81] -1.955102e-232 -6.522661e-234  5.369491e-234  4.396847e-233
#>   [85] -1.108431e-234  3.102998e-233 -3.154843e-233 -4.477472e-235
#>   [89]  5.793375e-234 -2.793062e-233  2.205193e-233 -3.462434e-233
#>   [93] -6.531600e-232  5.522789e-234  1.526689e-233 -3.383464e-234
#>   [97]  6.331418e-234 -5.570713e-234  1.271043e-232 -6.890614e-233
#>  [101] -1.151295e-233  9.363120e-235  7.300422e-233 -2.003917e-231
#>  [105] -1.674194e-233  3.004812e-233  3.537247e-231 -1.487991e-233
#>  [109] -4.913441e-233  5.558748e-234  1.876684e-233  3.432913e-234
#>  [113]  2.049151e-232 -5.614980e-233  1.751099e-235  4.854189e-233
#>  [117]  1.200734e-233 -2.212058e-234  1.049988e-233 -8.547264e-234
#>  [121]  6.221092e-233  3.730231e-234 -1.212666e-233  7.124796e-234
#>  [125] -1.381185e-232  7.270880e-234 -6.251382e-234 -1.020389e-232
#>  [129]  3.351242e-233  5.026236e-234  2.825432e-232  2.435571e-232
#>  [133]  4.538163e-234 -3.561182e-234 -3.366344e-231  2.694597e-234
#>  [137]  3.328102e-234  4.856633e-235 -2.629957e-233  4.903282e-233
#>  [141]  5.197222e-234 -4.727884e-234  8.765262e-234  2.924956e-234
#>  [145] -4.545786e-234 -8.817486e-234 -1.272293e-232  1.397281e-231
#>  [149]  9.838317e-234  3.832865e-233 -4.941206e-233  3.091898e-234
#>  [153]  1.608741e-233  2.351050e-233  5.037913e-234  1.871013e-234
#>  [157] -1.926804e-232 -1.109836e-233 -1.948927e-233 -1.118905e-232
#>  [161] -1.292627e-233 -8.468937e-234 -1.038218e-231 -1.532425e-231
#>  [165]  2.909617e-234 -2.987241e-233  7.445504e-232 -5.841763e-234
#>  [169]  1.104504e-234  3.698963e-233  1.723561e-231  1.769506e-233
#>  [173]  5.473197e-234 -2.949725e-233 -2.169850e-233 -4.951371e-231
#>  [177] -6.858440e-234 -1.881608e-233  2.069247e-232 -2.823617e-233
#>  [181]  7.099762e-232  1.085661e-234 -7.364690e-233 -1.604623e-233
#>  [185]  1.925845e-233  1.257426e-232  2.450071e-234 -3.705670e-232
#>  [189] -7.880927e-235 -1.419833e-233 -2.016980e-232 -3.437549e-234
#>  [193] -1.000287e-233  6.488986e-233  1.883741e-234  2.965577e-234
#>  [197] -1.492026e-233 -8.537341e-234 -2.486746e-233 -9.197567e-234
#>  [201]  1.101068e-233  4.987395e-233 -2.213851e-231  8.454600e-234
#>  [205] -3.807238e-232  3.588718e-232 -6.303156e-234  3.157200e-233
#>  [209]  1.068863e-232 -1.005560e-234 -2.158938e-232 -7.027313e-234
#>  [213]  7.237529e-235 -8.252201e-234 -1.556289e-230 -4.098915e-233
#>  [217] -1.546546e-234 -1.845707e-231 -4.249237e-233  1.270221e-232
#>  [221] -1.398645e-234 -6.975381e-234 -7.276006e-234 -6.525352e-234
#>  [225]  5.893430e-234 -5.590532e-233  1.050468e-233  2.431685e-233
#>  [229] -1.461085e-232 -6.311646e-233 -5.601160e-230  3.551742e-233
#>  [233] -8.759703e-234  2.047265e-233  7.569042e-233  1.750184e-232
#>  [237] -3.100583e-233  1.739241e-234 -9.476431e-233  4.548954e-233
#>  [241] -1.843040e-232  5.629190e-234 -8.066786e-235  7.673183e-233
#>  [245]  4.030746e-233  6.636193e-234  3.597584e-232  3.721546e-234
#>  [249] -1.039188e-233  6.553325e-235  1.609919e-233 -4.545289e-234
#>  [253]  9.542773e-235 -1.080553e-233 -1.110526e-233 -4.667686e-234
#>  [257]  1.759061e-233  3.824835e-234 -6.834182e-234  4.086325e-234
#>  [261]  3.115341e-234 -9.203927e-234 -2.776543e-234 -5.230669e-234
#>  [265] -4.684218e-233  1.934398e-232  2.294372e-234 -4.904567e-233
#>  [269] -1.692387e-233  2.597232e-233  1.709611e-233 -1.304863e-233
#>  [273] -8.804235e-235 -1.713660e-229 -2.443102e-234  2.385179e-230
#>  [277] -1.225163e-233  1.123689e-233  1.016844e-235 -2.836305e-233
#>  [281]  2.233138e-233 -1.932176e-232 -8.093893e-233  7.807999e-234
#>  [285] -2.536715e-233  1.009100e-235 -4.719182e-235 -1.595112e-233
#>  [289]  2.020583e-234 -7.208153e-235  7.488417e-234 -6.259644e-235
#>  [293]  1.475812e-231 -3.655090e-234  2.419853e-234 -2.315813e-233
#>  [297]  4.191332e-233 -4.642641e-232 -7.745495e-233  4.326011e-234
#>  [301]  3.672585e-233 -1.386148e-235 -7.823001e-234  9.767052e-230
#>  [305] -9.292904e-234  5.702036e-234 -3.510301e-233  7.253034e-233
#>  [309] -2.341628e-233 -2.329421e-232  1.722251e-233  6.074331e-234
#>  [313] -6.687483e-236  8.974455e-234  1.007217e-233  3.876618e-233
#>  [317]  1.510664e-232 -3.748556e-233  3.134320e-233  3.229150e-229
#>  [321] -1.065513e-233  4.118673e-234 -9.025388e-233  2.200693e-235
#>  [325]  5.905320e-233 -5.456369e-234  2.785496e-232 -1.766235e-233
#>  [329]  5.160700e-234 -1.108691e-232 -1.311211e-233  1.759830e-234
#>  [333]  1.453716e-232  4.951935e-233  6.310766e-235  1.365117e-234
#>  [337]  5.870717e-232 -6.846522e-233  3.819154e-233  1.097927e-234
#>  [341] -9.511195e-234  4.474487e-233 -2.120417e-234  2.914384e-233
#>  [345] -5.630783e-233  7.010608e-234  1.269962e-231  4.044631e-233
#>  [349] -1.350721e-233  1.476112e-234  7.122714e-233  2.688426e-233
#>  [353] -3.160748e-233  5.646995e-233 -3.795364e-232  1.221886e-232
#>  [357] -4.754916e-234 -1.035794e-231 -9.206505e-234 -7.565508e-231
#>  [361] -9.498267e-235 -3.075279e-232  1.524341e-232  5.036057e-235
#>  [365] -2.201768e-231 -8.960623e-232 -2.494874e-233 -1.015959e-233
#>  [369]  1.109871e-233  3.176054e-232  4.225858e-234 -6.398639e-233
#>  [373] -5.990845e-235  7.115459e-234 -3.805099e-232  1.070049e-233
#>  [377]  5.050973e-236 -4.202120e-233  4.517403e-233 -2.324356e-233
#>  [381]  7.383753e-233 -4.699071e-234 -1.404112e-233  8.720467e-231
#>  [385]  2.603877e-234 -1.432262e-231  2.530773e-234 -9.296344e-234
#>  [389] -1.661852e-233 -7.981428e-235  3.551412e-234  1.078492e-231
#>  [393]  1.169929e-233 -6.484630e-234 -4.749952e-234  6.116185e-234
#>  [397] -4.594989e-233 -1.618294e-230  4.586168e-234  9.293210e-235
#>  [401]  1.433225e-232  1.380834e-233 -1.800048e-233  2.485512e-233
#>  [405] -1.358740e-232  2.828054e-234 -1.488275e-233 -3.134908e-233
#>  [409] -5.260213e-234 -5.355288e-232 -2.952519e-233 -3.082782e-233
#>  [413]  1.073160e-233  3.687390e-233  3.741982e-234 -7.875619e-232
#>  [417] -1.405942e-233 -3.178142e-233  1.254542e-232 -1.604893e-233
#>  [421] -7.919441e-231  5.933093e-233  2.546949e-233  2.460433e-233
#>  [425] -3.780248e-234 -3.067494e-233 -5.332905e-233  3.885604e-234
#>  [429] -1.571229e-232  3.570647e-233 -1.594071e-233 -3.125984e-234
#>  [433]  3.426776e-234  1.593787e-231  1.813005e-233  2.236403e-233
#>  [437] -1.641413e-233  1.919369e-233  8.501940e-234  4.976947e-234
#>  [441] -1.761580e-233  1.817808e-233  3.538306e-234 -1.624767e-233
#>  [445] -3.272407e-233 -9.045473e-233 -3.181118e-234 -2.148313e-234
#>  [449] -8.559954e-233  2.267954e-234  9.239019e-234 -4.035755e-231
#>  [453] -2.507946e-233  3.600247e-233  9.397403e-234  8.237131e-233
#>  [457]  1.545523e-233  4.010319e-232 -2.565355e-233  3.410511e-232
#>  [461] -3.106537e-233  8.752112e-233  6.083317e-233 -2.720305e-233
#>  [465]  1.354221e-233 -2.125369e-234  5.768702e-234 -4.320711e-234
#>  [469]  4.762663e-233  1.406089e-232  1.794122e-232 -6.352538e-233
#>  [473]  1.124003e-233 -2.308650e-233  5.281105e-235  3.444259e-233
#>  [477] -1.237170e-233 -3.354405e-233  1.382252e-232 -7.989904e-234
#>  [481]  1.542172e-233 -1.135071e-232 -4.289455e-232  9.903117e-233
#>  [485] -4.756950e-234  1.598678e-234 -1.492673e-230 -1.365821e-233
#>  [489] -6.318221e-234 -5.106505e-233 -3.002568e-232 -7.635500e-234
#>  [493] -4.579854e-234  1.460071e-234 -5.158764e-233  4.254194e-235
#>  [497] -1.092282e-233  1.114923e-233  7.711920e-234 -5.879777e-234
#>  [501]  9.164188e-233  2.478926e-234 -4.523811e-232  2.113171e-232
#>  [505]  3.077877e-234  4.025846e-234 -7.411623e-234 -1.305333e-234
#>  [509]  1.747763e-233 -4.044142e-233  3.539301e-233 -1.446131e-232
#>  [513] -5.005272e-232  1.724599e-233  9.905934e-235 -1.132292e-231
#>  [517] -1.657202e-232  2.532753e-233  1.394224e-233 -1.235095e-231
#>  [521] -6.322660e-234 -5.357975e-236 -3.447306e-233 -4.059109e-235
#>  [525]  3.616341e-235  9.118702e-234  1.935387e-231  3.537974e-234
#>  [529] -5.855344e-234  3.046572e-232 -2.261305e-233  6.280832e-233
#>  [533]  3.436942e-233 -4.802257e-231 -1.710071e-233  4.527437e-234
#>  [537]  1.125301e-233 -4.057714e-233  8.723942e-234 -3.167394e-233
#>  [541] -3.937483e-234 -1.347304e-235  2.057741e-233  3.221415e-235
#>  [545] -4.644341e-233  2.061515e-233 -9.537040e-234 -7.992568e-234
#>  [549]  4.690166e-233  1.512118e-230  1.507866e-233  1.064106e-233
#>  [553] -4.778738e-234 -8.027209e-234 -1.247400e-231 -2.019402e-228
#>  [557]  6.258842e-234  1.777891e-234  2.758235e-233  1.143031e-233
#>  [561]  3.339491e-233  6.539160e-234 -3.729860e-233  3.727789e-234
#>  [565]  6.242154e-234  3.632196e-231 -3.494004e-234 -3.918962e-233
#>  [569] -3.615056e-232  5.415474e-233  2.456270e-234 -2.460974e-232
#>  [573] -6.418698e-233 -7.504388e-234 -6.465073e-234  2.539280e-233
#>  [577]  3.278559e-234  3.562376e-233  1.867074e-233  2.182948e-233
#>  [581] -1.136878e-234  2.261435e-233  3.331300e-233 -1.915303e-233
#>  [585]  3.699836e-233 -2.001624e-233 -1.244474e-232 -2.040320e-234
#>  [589]  8.836973e-234  2.546056e-234  5.618131e-233  5.351145e-233
#>  [593] -3.337970e-233  3.271946e-233  4.868897e-231 -1.228071e-233
#>  [597] -2.728068e-233  1.858152e-233  1.200642e-234  1.237019e-233
#>  [601]  1.745525e-232 -1.003999e-234  8.489249e-234  1.865431e-232
#>  [605]  3.241215e-234  6.447423e-233  9.605209e-234  1.103592e-234
#>  [609] -1.304213e-232  3.105286e-234  1.502877e-232  2.441007e-234
#>  [613]  4.053856e-228 -4.345288e-234 -4.871398e-233 -8.224651e-235
#>  [617] -1.270348e-233 -5.286719e-234  2.308917e-232 -1.092006e-232
#>  [621]  3.581109e-234  9.022567e-233  7.741930e-234 -3.417281e-234
#>  [625] -1.161591e-231 -2.678874e-232 -1.774470e-232 -1.515607e-233
#>  [629]  5.347105e-233 -1.227042e-233 -9.264159e-235 -1.603358e-232
#>  [633]  4.349710e-234 -3.750943e-233 -5.483775e-234  1.079139e-233
#>  [637] -6.285614e-232  4.110740e-233 -9.668760e-234  2.147356e-234
#>  [641] -9.549210e-233  4.720218e-232  8.645609e-235 -5.965570e-230
#>  [645] -5.361943e-233  1.918876e-234 -4.277795e-234 -2.610887e-234
#>  [649] -1.233002e-233  2.290883e-233 -1.447451e-233  4.437171e-234
#>  [653]  3.634894e-234  2.202424e-233 -6.434389e-235  5.375635e-233
#>  [657] -7.255834e-234 -8.465140e-234  5.368720e-233 -1.705321e-234
#>  [661]  5.571200e-234  9.528840e-233  1.937645e-234 -7.294040e-232
#>  [665]  9.956808e-234  1.219422e-233  5.435767e-234 -2.684375e-233
#>  [669] -6.153434e-234 -7.735230e-234  1.065786e-233  1.179952e-232
#>  [673] -1.797147e-235  1.526139e-233 -2.412748e-234  2.743288e-234
#>  [677]  6.216503e-234  1.198034e-234 -9.711988e-234 -9.530329e-235
#>  [681]  5.643863e-233  5.631400e-234  5.726676e-233 -2.619406e-233
#>  [685] -6.125629e-234 -1.854087e-233 -2.131619e-230 -1.179265e-233
#>  [689] -1.465569e-235  1.562481e-230  1.322619e-233  2.628733e-232
#>  [693]  3.050258e-233  2.195649e-234  1.000718e-233 -3.792130e-234
#>  [697] -7.101501e-234 -4.978016e-235  1.766590e-234  2.488288e-233
#>  [701] -7.955887e-234 -1.445063e-232 -1.112280e-232 -1.553036e-233
#>  [705]  8.456000e-234  1.208441e-233  9.413208e-233  8.158111e-233
#>  [709]  7.202501e-234 -2.463172e-233 -3.034785e-232  6.202822e-232
#>  [713]  9.391693e-232  6.782547e-234 -3.075939e-233  1.511976e-233
#>  [717] -4.551306e-232  3.321928e-233 -1.327823e-233 -7.207764e-234
#>  [721] -6.274613e-233  4.398460e-234  9.362939e-234 -5.734960e-234
#>  [725]  5.872749e-232 -1.845635e-233  2.097251e-234 -8.802923e-232
#>  [729] -1.143877e-230  1.873109e-235  1.377554e-233 -1.514782e-233
#>  [733]  2.748368e-230 -1.246943e-233 -5.699576e-233 -5.912375e-233
#>  [737] -6.521977e-234 -3.173701e-233  6.402925e-233 -2.223443e-233
#>  [741]  1.245748e-233 -5.810133e-233 -4.597874e-234  3.405186e-234
#>  [745]  5.096168e-235  1.243698e-232 -2.850689e-232 -9.086736e-234
#>  [749] -4.270814e-233 -1.785681e-232  7.302321e-234  2.692663e-233
#>  [753] -5.283302e-233  4.561210e-234 -4.721637e-234 -2.493182e-234
#>  [757] -3.541070e-233 -2.736778e-232  2.978880e-231 -5.439663e-235
#>  [761]  1.890524e-232  1.581418e-234  1.039369e-233 -5.263850e-232
#>  [765]  1.734898e-233 -1.313598e-232  2.978437e-231  1.294315e-233
#>  [769]  4.668795e-233 -1.250574e-233  6.385083e-234  1.086655e-232
#>  [773]  2.447600e-233 -1.369804e-233  1.247287e-230 -2.383023e-234
#>  [777]  1.803461e-235  1.308956e-233 -2.194235e-233 -2.618736e-233
#>  [781]  2.033398e-233  5.615256e-234  9.998157e-234  2.401523e-233
#>  [785]  1.353658e-233 -1.915107e-232 -2.540316e-234  2.067558e-233
#>  [789]  4.738654e-234 -9.733234e-233  5.397350e-235 -2.219224e-233
#>  [793]  3.302047e-233 -4.283373e-231  7.775853e-235  9.159569e-233
#>  [797] -5.131912e-234 -6.166943e-233 -3.187360e-232 -9.018147e-232
#>  [801] -3.860500e-233  4.230422e-234  8.117309e-232 -1.099657e-233
#>  [805]  1.843219e-233 -2.379812e-234 -7.411410e-234  6.338078e-234
#>  [809]  3.463044e-234 -9.693458e-233  4.278347e-232  9.272645e-234
#>  [813]  1.127817e-232  5.466265e-233  1.380917e-233 -2.752444e-233
#>  [817] -2.498377e-234 -1.842684e-233  1.794051e-231  3.615277e-234
#>  [821] -2.889458e-232 -7.131335e-233 -1.004303e-232 -3.077068e-234
#>  [825] -3.562339e-233 -3.319538e-233 -3.234410e-233 -9.035376e-235
#>  [829] -4.752739e-234  7.924730e-235  4.526932e-234  5.363956e-233
#>  [833] -5.732542e-233  2.548619e-232  1.649033e-233  1.076542e-234
#>  [837]  3.002067e-234 -8.268232e-234 -8.477408e-234  5.128850e-233
#>  [841]  1.507111e-233 -2.144883e-233  2.306234e-233  1.968780e-233
#>  [845]  1.313198e-233  5.049576e-233  1.607677e-233 -2.301901e-234
#>  [849] -2.087300e-233  2.297267e-233  1.258205e-234  3.632638e-234
#>  [853]  3.058456e-232  1.518797e-234  1.967566e-233  4.533256e-234
#>  [857] -2.171754e-232  1.376724e-234  5.378765e-234  2.084202e-233
#>  [861] -7.647967e-233 -2.706824e-233 -1.960903e-234  3.084587e-232
#>  [865] -1.914345e-234  1.941282e-233  4.558047e-233 -1.166762e-232
#>  [869]  4.573431e-234 -1.552516e-233  4.658696e-233 -2.819897e-234
#>  [873] -2.444306e-233 -1.005201e-232 -1.938759e-233  1.123503e-233
#>  [877]  6.389373e-234  7.472710e-233  9.188802e-232  2.435050e-234
#>  [881] -1.570196e-234  3.917193e-233 -1.424222e-233  1.131145e-232
#>  [885]  3.787377e-234 -3.452700e-233 -1.162786e-234 -2.155959e-233
#>  [889] -1.093759e-234  3.374498e-234  6.908545e-234 -1.767520e-232
#>  [893]  5.874719e-234 -1.326855e-233 -1.739012e-234 -1.057231e-234
#>  [897]  1.641364e-233  5.393337e-232 -5.109239e-232 -3.440688e-233
#>  [901] -6.834407e-234  9.534208e-232  1.070214e-233  1.021556e-232
#>  [905]  4.866962e-233 -1.191239e-232 -6.703760e-232 -2.973345e-233
#>  [909]  4.721139e-233 -9.788636e-234  1.183180e-233 -4.015284e-234
#>  [913] -1.649450e-233 -1.332194e-233  9.228800e-233  9.925280e-230
#>  [917]  1.195984e-235 -3.498587e-234 -3.174994e-235 -8.606748e-234
#>  [921] -1.850674e-233 -1.997848e-233 -1.420347e-233  7.097697e-233
#>  [925]  2.673713e-234 -2.943182e-234  1.528749e-234 -7.467181e-235
#>  [929] -1.554081e-234 -3.597629e-234 -2.772542e-232  1.949997e-233
#>  [933] -3.832562e-233  5.198859e-233 -2.555471e-230  1.344313e-233
#>  [937]  9.149903e-236 -1.799496e-233 -1.424275e-233  2.362913e-233
#>  [941] -4.847695e-233 -3.273247e-234  4.966449e-233 -1.924098e-234
#>  [945]  1.123236e-233 -2.632183e-231 -3.624277e-233  3.933865e-232
#>  [949]  5.473090e-234  2.234459e-235  2.100340e-233 -4.373893e-234
#>  [953] -3.800057e-233  7.099825e-234 -1.003007e-233  9.477840e-234
#>  [957] -2.235467e-234 -6.894097e-234  2.691475e-235 -3.321244e-232
#>  [961] -4.418511e-233  1.511264e-232  2.005676e-233  1.864939e-233
#>  [965] -5.219487e-233 -3.750567e-232  9.057301e-235 -8.197514e-230
#>  [969]  1.489892e-234 -3.282409e-234 -3.389659e-234  4.038863e-234
#>  [973] -3.907299e-234  2.057674e-234 -3.708232e-232 -7.177475e-233
#>  [977] -1.581247e-233 -9.069109e-231  1.209852e-234  1.095763e-233
#>  [981]  1.540250e-232  5.781881e-234 -6.933718e-233  4.335351e-234
#>  [985] -1.486229e-232 -5.625135e-234  1.508889e-234 -9.058345e-234
#>  [989] -5.809560e-234 -1.639880e-233 -7.230293e-233 -1.111586e-233
#>  [993]  1.443371e-233 -3.679926e-231 -1.588406e-232  7.726596e-235
#>  [997]  9.298383e-231 -5.963830e-234  2.213059e-233 -1.051725e-233
#> [1001]  1.140135e-231

res2      <- with(ss,susie_suff_stat(bhat = betahat, shat = sebetahat, R = R, n=n, var_y = var(y)))
coef(res2)

#>    [1]   0.000000e+00   1.061698e+00   1.023098e+00   1.004091e+00
#>    [5]   9.891083e-01 -7.784212e-234 -1.761841e-230  9.977581e-232
#>    [9]  4.036125e-233  1.419073e-233 -4.118527e-234  1.834910e-233
#>   [13]  4.586684e-233  9.731062e-234  2.256263e-233 -1.041226e-233
#>   [17]  3.521439e-235  9.605938e-233  2.633381e-232  2.389299e-233
#>   [21] -3.047347e-233 -4.659701e-232 -1.568859e-232  2.216254e-232
#>   [25] -3.883943e-233  4.299220e-233 -1.199493e-233  8.675355e-234
#>   [29]  1.572896e-234  4.353059e-234 -1.025333e-233 -1.937613e-234
#>   [33]  8.827845e-231  6.316872e-234 -5.520344e-234 -3.807140e-234
#>   [37]  1.492908e-229  3.186259e-234 -3.282107e-233 -1.347556e-233
#>   [41]  4.008097e-233  4.333559e-234  6.700667e-235 -7.722894e-234
#>   [45] -7.684455e-233 -4.189512e-233  1.625133e-230 -6.517148e-232
#>   [49]  1.830235e-233 -2.083339e-234  3.497915e-232 -1.309959e-232
#>   [53] -2.443137e-234  5.364331e-235 -8.672070e-234  1.192368e-233
#>   [57]  4.460670e-233  5.770805e-234 -4.751142e-234 -5.755387e-233
#>   [61]  1.148279e-232  5.609641e-234 -2.989580e-232  3.156000e-234
#>   [65] -5.286838e-234 -2.401684e-234  4.803885e-234  1.121453e-232
#>   [69]  1.816221e-233 -1.266846e-229  4.532927e-234  6.975123e-233
#>   [73]  1.778310e-231  2.281354e-234 -5.938679e-232 -3.090638e-233
#>   [77] -3.352356e-234  9.516434e-234  1.320660e-232  4.184417e-231
#>   [81] -1.943731e-232 -6.591258e-234  5.513514e-234  4.309800e-233
#>   [85] -1.092085e-234  3.125828e-233 -3.132561e-233 -4.554983e-235
#>   [89]  5.947470e-234 -2.771614e-233  2.116596e-233 -3.453008e-233
#>   [93] -6.542736e-232  5.342118e-234  1.528111e-233 -3.300944e-234
#>   [97]  6.332019e-234 -5.669709e-234  1.278234e-232 -6.741071e-233
#>  [101] -1.103262e-233  9.494402e-235  7.100625e-233 -1.966818e-231
#>  [105] -1.726735e-233  3.134710e-233  3.529717e-231 -1.576304e-233
#>  [109] -5.092025e-233  5.527054e-234  1.813597e-233  3.377802e-234
#>  [113]  2.118444e-232 -5.401874e-233  1.741359e-235  4.860250e-233
#>  [117]  1.183318e-233 -2.216566e-234  1.052142e-233 -8.652038e-234
#>  [121]  6.231378e-233  3.830989e-234 -1.195317e-233  7.101318e-234
#>  [125] -1.387745e-232  7.057346e-234 -6.715374e-234 -9.986419e-233
#>  [129]  3.313643e-233  4.865494e-234  2.801283e-232  2.339200e-232
#>  [133]  4.578015e-234 -3.521447e-234 -3.253012e-231  2.682411e-234
#>  [137]  3.322879e-234  4.860001e-235 -2.605231e-233  4.896684e-233
#>  [141]  5.172439e-234 -4.678189e-234  8.610755e-234  2.927584e-234
#>  [145] -4.425314e-234 -8.775174e-234 -1.311007e-232  1.356928e-231
#>  [149]  9.628683e-234  3.812873e-233 -4.956342e-233  3.050212e-234
#>  [153]  1.551638e-233  2.319946e-233  4.983849e-234  1.831653e-234
#>  [157] -1.847673e-232 -1.140504e-233 -1.931587e-233 -1.092240e-232
#>  [161] -1.312559e-233 -8.632974e-234 -1.040154e-231 -1.584082e-231
#>  [165]  2.873440e-234 -3.035816e-233  7.413556e-232 -6.041538e-234
#>  [169]  1.098205e-234  3.926899e-233  1.738529e-231  1.737964e-233
#>  [173]  5.738618e-234 -2.909433e-233 -2.124178e-233 -4.906229e-231
#>  [177] -6.852830e-234 -1.900137e-233  2.164834e-232 -2.765999e-233
#>  [181]  6.991977e-232  1.116591e-234 -7.547989e-233 -1.622347e-233
#>  [185]  1.841419e-233  1.293102e-232  2.477266e-234 -3.677488e-232
#>  [189] -7.842462e-235 -1.422271e-233 -1.987584e-232 -3.310976e-234
#>  [193] -1.013991e-233  6.592945e-233  1.849319e-234  3.046769e-234
#>  [197] -1.476331e-233 -8.290497e-234 -2.659168e-233 -9.030297e-234
#>  [201]  1.095561e-233  5.056066e-233 -2.236355e-231  8.174683e-234
#>  [205] -3.837447e-232  3.677481e-232 -6.567643e-234  3.191643e-233
#>  [209]  1.111691e-232 -9.734910e-235 -2.118431e-232 -6.667209e-234
#>  [213]  7.336086e-235 -8.175773e-234 -1.570093e-230 -4.112885e-233
#>  [217] -1.544442e-234 -1.830718e-231 -4.299234e-233  1.313200e-232
#>  [221] -1.437004e-234 -7.193280e-234 -7.115129e-234 -6.599219e-234
#>  [225]  5.838623e-234 -5.865453e-233  1.069085e-233  2.487339e-233
#>  [229] -1.493477e-232 -6.351685e-233 -5.590830e-230  3.516094e-233
#>  [233] -9.104065e-234  2.067720e-233  7.514766e-233  1.813112e-232
#>  [237] -3.125821e-233  1.693595e-234 -9.295029e-233  4.595188e-233
#>  [241] -1.792328e-232  5.717357e-234 -8.072396e-235  7.840328e-233
#>  [245]  3.951924e-233  6.697473e-234  3.579402e-232  3.636037e-234
#>  [249] -1.071123e-233  6.510292e-235  1.550329e-233 -4.526795e-234
#>  [253]  9.711758e-235 -1.073895e-233 -1.122623e-233 -4.713019e-234
#>  [257]  1.844245e-233  3.735848e-234 -6.814602e-234  4.025176e-234
#>  [261]  3.110022e-234 -9.016556e-234 -2.649379e-234 -5.114751e-234
#>  [265] -4.813099e-233  1.920451e-232  2.266718e-234 -5.010076e-233
#>  [269] -1.725825e-233  2.702517e-233  1.762438e-233 -1.359627e-233
#>  [273] -8.936491e-235 -1.684837e-229 -2.366813e-234  2.354185e-230
#>  [277] -1.259502e-233  1.119647e-233  1.002141e-235 -2.737898e-233
#>  [281]  2.199827e-233 -1.904571e-232 -8.193952e-233  7.976417e-234
#>  [285] -2.509897e-233  1.009057e-235 -4.789224e-235 -1.601282e-233
#>  [289]  2.007598e-234 -7.043500e-235  7.544328e-234 -6.301872e-235
#>  [293]  1.495086e-231 -3.653898e-234  2.458276e-234 -2.323961e-233
#>  [297]  4.296998e-233 -4.594747e-232 -8.010143e-233  4.173914e-234
#>  [301]  3.794636e-233 -1.369367e-235 -7.878801e-234  9.413530e-230
#>  [305] -9.293578e-234  5.990512e-234 -3.476265e-233  7.108338e-233
#>  [309] -2.288333e-233 -2.412656e-232  1.642675e-233  6.078194e-234
#>  [313] -6.815714e-236  8.687279e-234  9.863585e-234  3.806781e-233
#>  [317]  1.486057e-232 -3.718020e-233  3.056210e-233  3.159952e-229
#>  [321] -1.065627e-233  4.106290e-234 -8.811021e-233  2.196648e-235
#>  [325]  5.753034e-233 -5.412095e-234  2.689181e-232 -1.806522e-233
#>  [329]  5.306094e-234 -1.146140e-232 -1.313416e-233  1.689730e-234
#>  [333]  1.492795e-232  4.974064e-233  6.205538e-235  1.365842e-234
#>  [337]  5.595775e-232 -6.654305e-233  3.859684e-233  1.078670e-234
#>  [341] -9.369084e-234  4.470505e-233 -2.157038e-234  2.872088e-233
#>  [345] -5.627704e-233  7.312753e-234  1.308860e-231  4.058671e-233
#>  [349] -1.344001e-233  1.473956e-234  7.200071e-233  2.836847e-233
#>  [353] -3.273720e-233  5.807096e-233 -3.901367e-232  1.303795e-232
#>  [357] -4.784240e-234 -9.777495e-232 -9.141687e-234 -7.408473e-231
#>  [361] -9.260174e-235 -3.139546e-232  1.549934e-232  5.167518e-235
#>  [365] -2.187000e-231 -9.212429e-232 -2.499554e-233 -1.044271e-233
#>  [369]  1.137444e-233  3.208446e-232  4.153223e-234 -6.567913e-233
#>  [373] -6.262807e-235  6.960097e-234 -3.815166e-232  1.062505e-233
#>  [377]  5.255794e-236 -4.187352e-233  4.493734e-233 -2.359848e-233
#>  [381]  7.269667e-233 -4.797665e-234 -1.393664e-233  8.764488e-231
#>  [385]  2.538396e-234 -1.448903e-231  2.501798e-234 -9.111988e-234
#>  [389] -1.696052e-233 -8.202315e-235  3.604865e-234  1.073775e-231
#>  [393]  1.190989e-233 -6.579709e-234 -4.758608e-234  6.104339e-234
#>  [397] -4.641535e-233 -1.653365e-230  4.547988e-234  9.000718e-235
#>  [401]  1.455585e-232  1.434205e-233 -1.850212e-233  2.448367e-233
#>  [405] -1.335710e-232  2.839275e-234 -1.544156e-233 -3.074300e-233
#>  [409] -5.282581e-234 -5.372329e-232 -2.843312e-233 -3.123780e-233
#>  [413]  1.090113e-233  3.607949e-233  3.640557e-234 -8.047676e-232
#>  [417] -1.403010e-233 -3.173987e-233  1.236384e-232 -1.647054e-233
#>  [421] -7.841483e-231  5.989267e-233  2.435490e-233  2.475259e-233
#>  [425] -3.683001e-234 -3.157886e-233 -5.262971e-233  3.977658e-234
#>  [429] -1.543278e-232  3.584409e-233 -1.612150e-233 -3.185603e-234
#>  [433]  3.321776e-234  1.591996e-231  1.800769e-233  2.230896e-233
#>  [437] -1.623740e-233  1.890100e-233  8.659630e-234  4.876865e-234
#>  [441] -1.780617e-233  1.816876e-233  3.407773e-234 -1.694600e-233
#>  [445] -3.239333e-233 -8.561217e-233 -3.303366e-234 -2.182529e-234
#>  [449] -8.624069e-233  2.232362e-234  8.810628e-234 -4.185730e-231
#>  [453] -2.478347e-233  3.603149e-233  9.629157e-234  8.316867e-233
#>  [457]  1.587076e-233  4.098613e-232 -2.552032e-233  3.239854e-232
#>  [461] -3.135554e-233  8.859510e-233  6.128506e-233 -2.720515e-233
#>  [465]  1.360443e-233 -2.132414e-234  6.147171e-234 -4.263897e-234
#>  [469]  4.703698e-233  1.420221e-232  1.784671e-232 -6.315772e-233
#>  [473]  1.121406e-233 -2.357356e-233  5.146493e-235  3.524421e-233
#>  [477] -1.226646e-233 -3.385474e-233  1.420733e-232 -7.916011e-234
#>  [481]  1.513662e-233 -1.099046e-232 -4.423134e-232  9.917621e-233
#>  [485] -4.781999e-234  1.604545e-234 -1.544074e-230 -1.432806e-233
#>  [489] -6.316906e-234 -5.070384e-233 -2.982381e-232 -7.522274e-234
#>  [493] -4.485459e-234  1.470673e-234 -5.137277e-233  4.377406e-235
#>  [497] -1.119098e-233  1.121021e-233  7.467966e-234 -6.032170e-234
#>  [501]  9.529950e-233  2.530223e-234 -4.415953e-232  2.099981e-232
#>  [505]  3.211498e-234  3.925803e-234 -7.349692e-234 -1.289919e-234
#>  [509]  1.743279e-233 -4.023029e-233  3.590916e-233 -1.443499e-232
#>  [513] -5.128128e-232  1.706377e-233  9.533067e-235 -1.131784e-231
#>  [517] -1.714720e-232  2.548588e-233  1.396485e-233 -1.245804e-231
#>  [521] -6.525550e-234 -5.437902e-236 -3.352819e-233 -4.000013e-235
#>  [525]  3.578219e-235  9.275070e-234  1.947182e-231  3.518522e-234
#>  [529] -5.806798e-234  3.082339e-232 -2.209189e-233  6.100520e-233
#>  [533]  3.483631e-233 -4.887081e-231 -1.721389e-233  4.516220e-234
#>  [537]  1.146151e-233 -4.089910e-233  8.691995e-234 -3.294729e-233
#>  [541] -3.911538e-234 -1.372464e-235  1.987645e-233  3.233291e-235
#>  [545] -4.492453e-233  2.110313e-233 -9.568577e-234 -7.908299e-234
#>  [549]  4.987631e-233  1.499114e-230  1.493964e-233  1.075207e-233
#>  [553] -5.013381e-234 -7.953031e-234 -1.246424e-231 -2.042621e-228
#>  [557]  6.136656e-234  1.787281e-234  2.617942e-233  1.159884e-233
#>  [561]  3.350417e-233  6.222871e-234 -3.896693e-233  3.687017e-234
#>  [565]  6.240447e-234  3.654743e-231 -3.516334e-234 -4.008831e-233
#>  [569] -3.744693e-232  5.463984e-233  2.417335e-234 -2.411183e-232
#>  [573] -6.224383e-233 -7.521710e-234 -6.693166e-234  2.539623e-233
#>  [577]  3.205053e-234  3.550667e-233  1.882472e-233  2.152541e-233
#>  [581] -1.127095e-234  2.236513e-233  3.380998e-233 -1.898530e-233
#>  [585]  3.487985e-233 -1.958458e-233 -1.216031e-232 -2.093050e-234
#>  [589]  8.484678e-234  2.503475e-234  5.686412e-233  5.328359e-233
#>  [593] -3.332200e-233  3.311312e-233  4.847643e-231 -1.264543e-233
#>  [597] -2.841747e-233  1.837026e-233  1.237230e-234  1.235126e-233
#>  [601]  1.705887e-232 -1.021264e-234  8.275200e-234  1.901746e-232
#>  [605]  3.244724e-234  6.626402e-233  9.836350e-234  1.083893e-234
#>  [609] -1.384107e-232  3.155688e-234  1.457250e-232  2.381059e-234
#>  [613]  4.069980e-228 -4.214845e-234 -4.667535e-233 -8.105345e-235
#>  [617] -1.285707e-233 -5.391525e-234  2.366586e-232 -1.092896e-232
#>  [621]  3.643436e-234  9.090509e-233  7.464751e-234 -3.510394e-234
#>  [625] -1.170594e-231 -2.777365e-232 -1.709352e-232 -1.502808e-233
#>  [629]  5.163379e-233 -1.244076e-233 -9.150559e-235 -1.581772e-232
#>  [633]  4.274055e-234 -3.879986e-233 -5.342090e-234  1.079853e-233
#>  [637] -6.345184e-232  4.302551e-233 -9.768776e-234  2.192224e-234
#>  [641] -9.805494e-233  5.064797e-232  8.654220e-235 -5.823762e-230
#>  [645] -5.281337e-233  1.837126e-234 -4.216164e-234 -2.662247e-234
#>  [649] -1.227146e-233  2.350366e-233 -1.430152e-233  4.341746e-234
#>  [653]  3.596100e-234  2.168922e-233 -6.362223e-235  5.534590e-233
#>  [657] -7.439412e-234 -8.830421e-234  5.405847e-233 -1.731826e-234
#>  [661]  5.607448e-234  9.561756e-233  2.042030e-234 -7.183594e-232
#>  [665]  9.993667e-234  1.200098e-233  5.634165e-234 -2.656416e-233
#>  [669] -6.224432e-234 -7.476801e-234  1.042266e-233  1.171458e-232
#>  [673] -1.761848e-235  1.541056e-233 -2.433708e-234  2.712696e-234
#>  [677]  6.104578e-234  1.222324e-234 -9.601533e-234 -9.605643e-235
#>  [681]  5.727658e-233  5.405467e-234  5.774797e-233 -2.607099e-233
#>  [685] -6.118535e-234 -1.892703e-233 -2.133658e-230 -1.159292e-233
#>  [689] -1.464129e-235  1.584811e-230  1.316033e-233  2.646454e-232
#>  [693]  3.078953e-233  2.200757e-234  1.013307e-233 -3.702094e-234
#>  [697] -7.348981e-234 -5.059413e-235  1.775595e-234  2.504605e-233
#>  [701] -7.968486e-234 -1.413281e-232 -1.121623e-232 -1.592027e-233
#>  [705]  8.290923e-234  1.183092e-233  9.303039e-233  8.141904e-233
#>  [709]  7.330952e-234 -2.503105e-233 -3.072087e-232  6.083712e-232
#>  [713]  9.198210e-232  6.769237e-234 -3.206887e-233  1.500897e-233
#>  [717] -4.572827e-232  3.353404e-233 -1.295247e-233 -7.373772e-234
#>  [721] -6.359883e-233  4.415930e-234  9.397176e-234 -5.581657e-234
#>  [725]  6.129850e-232 -1.829906e-233  2.110539e-234 -8.767028e-232
#>  [729] -1.106531e-230  1.852707e-235  1.376914e-233 -1.579115e-233
#>  [733]  2.671736e-230 -1.232507e-233 -5.694552e-233 -5.844860e-233
#>  [737] -6.575242e-234 -3.096255e-233  6.329998e-233 -2.254020e-233
#>  [741]  1.267077e-233 -5.780490e-233 -4.575986e-234  3.338938e-234
#>  [745]  5.051151e-235  1.241348e-232 -2.859173e-232 -9.096201e-234
#>  [749] -4.155655e-233 -1.824978e-232  7.296422e-234  2.723516e-233
#>  [753] -5.149027e-233  4.730113e-234 -4.520574e-234 -2.453343e-234
#>  [757] -3.623705e-233 -2.734029e-232  2.874599e-231 -5.560453e-235
#>  [761]  1.837669e-232  1.586651e-234  1.060747e-233 -5.262321e-232
#>  [765]  1.721020e-233 -1.343502e-232  3.062493e-231  1.281179e-233
#>  [769]  4.558084e-233 -1.249473e-233  6.532514e-234  1.087762e-232
#>  [773]  2.322377e-233 -1.340564e-233  1.259610e-230 -2.386197e-234
#>  [777]  1.804782e-235  1.318444e-233 -2.204101e-233 -2.638071e-233
#>  [781]  2.054878e-233  5.630487e-234  9.833317e-234  2.472686e-233
#>  [785]  1.344712e-233 -1.951738e-232 -2.514044e-234  2.064889e-233
#>  [789]  4.688543e-234 -9.784806e-233  5.305985e-235 -2.170758e-233
#>  [793]  3.416884e-233 -4.366604e-231  7.719209e-235  9.420795e-233
#>  [797] -4.991276e-234 -6.032207e-233 -3.154090e-232 -9.007008e-232
#>  [801] -3.996144e-233  4.388721e-234  8.319092e-232 -1.089016e-233
#>  [805]  1.902948e-233 -2.365662e-234 -7.242335e-234  6.269446e-234
#>  [809]  3.405658e-234 -9.992144e-233  4.259930e-232  8.979833e-234
#>  [813]  1.117903e-232  5.514558e-233  1.363129e-233 -2.698575e-233
#>  [817] -2.540046e-234 -1.804669e-233  1.766075e-231  3.638196e-234
#>  [821] -2.988091e-232 -7.145474e-233 -1.024582e-232 -3.110737e-234
#>  [825] -3.470555e-233 -3.371517e-233 -3.212399e-233 -8.910009e-235
#>  [829] -4.906368e-234  7.920659e-235  4.571855e-234  5.344932e-233
#>  [833] -5.935452e-233  2.522057e-232  1.644564e-233  1.022688e-234
#>  [837]  3.114984e-234 -8.474212e-234 -8.213995e-234  5.045188e-233
#>  [841]  1.512233e-233 -2.069552e-233  2.310448e-233  2.023409e-233
#>  [845]  1.257670e-233  4.931484e-233  1.585173e-233 -2.246596e-234
#>  [849] -2.135844e-233  2.253114e-233  1.256507e-234  3.718080e-234
#>  [853]  3.138628e-232  1.533768e-234  2.014741e-233  4.595313e-234
#>  [857] -2.209902e-232  1.399670e-234  5.622696e-234  2.099851e-233
#>  [861] -7.697572e-233 -2.665030e-233 -1.971928e-234  2.998055e-232
#>  [865] -2.008669e-234  1.933087e-233  4.459109e-233 -1.142318e-232
#>  [869]  4.469615e-234 -1.579381e-233  4.702264e-233 -2.738116e-234
#>  [873] -2.449981e-233 -1.031003e-232 -1.935567e-233  1.098577e-233
#>  [877]  6.443507e-234  7.470824e-233  9.038690e-232  2.423643e-234
#>  [881] -1.531266e-234  3.837753e-233 -1.511308e-233  1.157349e-232
#>  [885]  3.695803e-234 -3.574843e-233 -1.209586e-234 -2.177079e-233
#>  [889] -1.060589e-234  3.268720e-234  6.538199e-234 -1.819186e-232
#>  [893]  5.791350e-234 -1.310157e-233 -1.803298e-234 -1.052049e-234
#>  [897]  1.621126e-233  5.500994e-232 -5.214999e-232 -3.363229e-233
#>  [901] -7.070959e-234  9.846192e-232  1.069728e-233  9.887105e-233
#>  [905]  4.714879e-233 -1.219970e-232 -6.420227e-232 -2.995163e-233
#>  [909]  4.793432e-233 -9.998079e-234  1.236714e-233 -4.128535e-234
#>  [913] -1.680407e-233 -1.302931e-233  9.111917e-233  1.022526e-229
#>  [917]  1.219242e-235 -3.354234e-234 -3.127361e-235 -9.073060e-234
#>  [921] -1.899672e-233 -1.991294e-233 -1.358381e-233  7.168816e-233
#>  [925]  2.701837e-234 -2.964705e-234  1.561135e-234 -7.589846e-235
#>  [929] -1.528326e-234 -3.496230e-234 -2.832906e-232  1.913902e-233
#>  [933] -3.844030e-233  5.197205e-233 -2.466558e-230  1.392169e-233
#>  [937]  9.430437e-236 -1.782377e-233 -1.398322e-233  2.286184e-233
#>  [941] -4.872861e-233 -3.244625e-234  4.940242e-233 -1.919637e-234
#>  [945]  1.097881e-233 -2.517569e-231 -3.785280e-233  3.919005e-232
#>  [949]  5.854513e-234  2.231941e-235  2.097292e-233 -4.238844e-234
#>  [953] -3.765613e-233  6.937094e-234 -1.013375e-233  9.387649e-234
#>  [957] -2.211942e-234 -6.836297e-234  2.699611e-235 -3.360839e-232
#>  [961] -4.223336e-233  1.496284e-232  1.939069e-233  1.830518e-233
#>  [965] -5.210968e-233 -3.957936e-232  9.099570e-235 -8.108475e-230
#>  [969]  1.459099e-234 -3.259142e-234 -3.422796e-234  3.936342e-234
#>  [973] -3.898357e-234  2.040354e-234 -3.709337e-232 -7.061669e-233
#>  [977] -1.541460e-233 -8.997651e-231  1.228201e-234  1.083683e-233
#>  [981]  1.522343e-232  5.700257e-234 -7.158561e-233  4.449673e-234
#>  [985] -1.505874e-232 -5.621286e-234  1.486357e-234 -8.918970e-234
#>  [989] -5.972010e-234 -1.677798e-233 -7.059244e-233 -1.092667e-233
#>  [993]  1.391815e-233 -3.686532e-231 -1.575360e-232  7.594912e-235
#>  [997]  9.309125e-231 -6.094262e-234  2.231241e-233 -1.008467e-233
#> [1001]  1.125563e-231