Summary statistics version of SuSiE on z scores and correlation (or covariance) matrix.

Summary statistics version of SuSiE on z scores and correlation (or covariance) matrix.

susie_z(z, R, r_tol = 1e-08, L = 10,
  estimate_residual_variance = TRUE, estimate_prior_method = c("optim",
  "EM"), prior_weights = NULL, null_weight = NULL, coverage = 0.95,
  min_abs_corr = 0.5, verbose = FALSE, track_fit = FALSE, ...)

Arguments

z
a p vector of z scores.

R
a p by p symmetric and positive semidefinite matrix. It can be X'X, covariance matrix or correlation matrix.

r_tol
tolerance level for eigen value check of positive semidefinite matrix of R.

L
maximum number of non-zero effects.

estimate_residual_variance
indicates whether to estimate residual variance

estimate_prior_method
The method used for estimating prior variance, 'optim' or 'EM'

prior_weights
a p vector of prior probability that each element is non-zero.

null_weight
probability of no effect, for each single effect model.

coverage
coverage of confident sets. Default to 0.95 for 95% credible interval.

min_abs_corr
minimum of absolute value of correlation allowed in a credible set.

verbose
if TRUE outputs some progress messages.

track_fit
add an attribute trace to output that saves current values of all iterations.

...
further arguments to be passed to susie_ss

Value

a susie fit