Bayesian sum of single-effect (susie) linear regression of Y on X

susie(X, Y, L = 10, prior_variance = 0.2, residual_variance = NULL,
  standardize = TRUE, intercept = TRUE, max_iter = 100, tol = 0.01,
  estimate_residual_variance = TRUE, estimate_prior_variance = FALSE,
  s_init = NULL, verbose = FALSE)

Arguments

X
an n by p matrix of covariates

Y
an n vector

L
maximum number of non-zero effects

prior_variance
the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times )

residual_variance
the residual variance (defaults to variance of Y)

standardize
logical flag for whether to standardize columns of X to unit variance prior to fitting. Note that `prior_variance` specifies the prior on the coefficients of X after standardization (if performed). If you do not standardize you may need to think carefully about specifying `prior_variance`. Whatever the value of standardize, the coefficients (returned from `coef`) are for X on the original input scale. Any column of X that has zero variance is not standardized, but left as is.

intercept
Should intercept be fitted (default=TRUE) or set to zero (FALSE)

max_iter
maximum number of iterations to perform

tol
convergence tolerance

estimate_residual_variance
indicates whether to estimate residual variance

estimate_prior_variance
indicates whether to estimate prior (currently not recommended as not working as well)

s_init
a previous susie fit with which to initialize

verbose
if true outputs some progress messages

Value

a susie fit, which is a list with some or all of the following elements

alpha

an L by p matrix of posterior inclusion probabilites

mu

an L by p matrix of posterior means (conditional on inclusion)

mu2

an L by p matrix of posterior second moments (conditional on inclusion)

Xr

an n vector of fitted values, equal to X times colSums(alpha*mu))

sigma2

residual variance

sa2

scaled prior variance; ie prior variance is sigma2*sa2

elbo

vector of values of elbo achieved (objective function)

Details

Performs sum of single-effect (susie) linear regression of Y on X. That is, this function fits the regression model Y= sum_l Xb_l + e, where elements of e are iid N(0,s2) and the sum_l b_l is a p vector of effects to be estimated. The assumption is that each b_l has exactly one non-zero element, with all elements equally likely to be non-zero. The prior on the non-zero element is N(0,var=sa2*s2).

Examples

# NOT RUN {
set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1] = 1
beta[2] = 1
beta[3] = 1
beta[4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = X %*% beta + rnorm(n)
res =susie(X,y,L=10)
coef(res)
plot(y,predict(res))
# }