SUm of Single Effects (SuSiE) Regression

Performs Bayesian multiple linear regression of Y on X; that is, this function fits the regression model \(Y = sum_l Xb_l + e\), where elements of e are i.i.d. normal with zero mean and variance residual_variance, and \(sum_l b_l\) is a vector of length p representing the effects to be estimated. The SuSiE assumption is that each b_l has exactly one non-zero element. The prior on the non-zero element is normal with zero mean and variance var(Y)*scaled_prior_variance.

The model is fitted using the "Iterative Bayesian Stepwise Selection" (IBSS) algorithm.

susie(X, Y, L = min(10, ncol(X)), scaled_prior_variance = 0.2,
  residual_variance = NULL, prior_weights = NULL, null_weight = NULL,
  standardize = TRUE, intercept = TRUE,
  estimate_residual_variance = TRUE, estimate_prior_variance = TRUE,
  estimate_prior_method = c("optim", "EM"), s_init = NULL,
  coverage = 0.95, min_abs_corr = 0.5,
  compute_univariate_zscore = FALSE, max_iter = 100, tol = 0.001,
  verbose = FALSE, track_fit = FALSE)

Arguments

X
An n by p matrix of covariates.

Y
A vector of length n.

L
Number of components (nonzero elements) in the SuSiE regression model. If L is larger than the number of covariate (p), L is set to p.

scaled_prior_variance
The scaled prior variance. This is either a scalar, or a vector of length L. The prior variance of each non-zero element of b is set to var(Y)*scaled_prior_variance. If estimate_prior_variance = TRUE, this input provides the initial estimates of the prior variances.

residual_variance
The variance of the residual. If estimate_residual_variance = TRUE, this value provides the initial estimate of the residual variance. By default, it is set to var(Y).

prior_weights
A vector of length p, in which each entry gives the prior probability that corresponding column of X has a nonzero effect on the outcome, Y.

null_weight
Prior probability of no effect (a number between 0 and 1, and cannot be exactly 1).

standardize
If standardize = TRUE, standardize the columns of X to unit variance prior to fitting. Note that `scaled_prior_variance` specifies the prior on the coefficients of X after standardization (if it is performed). If you do not standardize, you may need to think more carefully about specifying scaled_prior_variance. Whatever your choice, the coefficients returned by coef are given for X on the original input scale. Any column of X that has zero variance is not standardized, but left as is.

intercept
If intercept = TRUE, the intercept is fitted; otherwise, it is set to zero. Setting intercept = FALSE is generally not recommended.

estimate_residual_variance
If estimate_residual_variance = TRUE, the variance of the residual is estimated separately for each of the L components, and scaled_prior_variance is used as an initial estimate of the variances.

estimate_prior_variance
If estimate_prior_variance = TRUE, the prior variance is estimated father than fixed.

estimate_prior_method
The method used for estimating prior variance.

s_init
A previous susie fit with which to initialize.

coverage
A number between 0 and 1 specifying the coverage of the estimated confidence sets.

min_abs_corr
Minimum of absolute value of correlation allowed in a credible set. The default, 0.5, corresponds to squared correlation of 0.25, which is a commonly used threshold for genotype data in genetics studies.

compute_univariate_zscore
If compute_univariate_zscore = TRUE, the univariate regression z-scores are outputted for each variable.

max_iter
Maximum number of iterations of the IBSS fitting procedure.

tol
A small, non-negative number specifying the convergence tolerance for the IBSS fitting procedure. The fitting procedure will halt when the difference in the variational lower bound, or "ELBO" (this is the objective function to be maximized), is less than tol.

verbose
If verbose = TRUE, the algorithm's progress and a summary of the optimization settings are printed to the console.

track_fit
If track_fit = TRUE, an object trace is also returned containing detailed information about the estimates at each iteration of the IBSS fitting procedure.

Value

A "susie" object with some or all of the following elements:

alpha

An L by p matrix of posterior inclusion probabilites.

mu

An L by p matrix of posterior means, conditional on inclusion.

mu2

An L by p matrix of posterior second moments, conditional on inclusion.

Xr

An vector of length n, equal to X %*% colSums(alpha * mu).

intercept

The intercept (fixed or estimated).

sigma2

Residual variance (fixed or estimated).

V

Prior variance of the non-zero elements of b, equal to scaled_prior_variance * var(Y).

elbo

The value of the variational lower bound, or "ELBO" (the objective function to be maximized), achieved at each iteration of the IBSS fitting procedure.

fitted

Vector of length n containing the "fitted" values of the outcome.

sets

Credible sets estimated from model fit; see susie_get_cs for details.

pip

A vector of length p giving the (marginal) posterior inclusion probabilities for all p covariates.

z

A vector of univariate z-scores.

niter

Number of IBSS iterations that were run.

converged

TRUE or FALSE, indicating whether the IBSS converged to a solution within the chosen tolerance level.

References

G. Wang, A. Sarkar, P. Carbonetto and M. Stephens (2018). A simple new approach to variable selection in regression, with application to genetic fine-mapping. bioRxiv https://doi.org/10.1101/501114.

Examples

# NOT RUN {
set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1:4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = X %*% beta + rnorm(n)
res = susie(X,y,L=10)
coef(res)
plot(y,predict(res))

# }