Inferences From Fitted SuSiE Model

These functions access basic properties or draw inferences from a fitted susie model.

Usage

susie_get_objective(res, last_only = TRUE, warning_tol = 1e-06)

susie_get_posterior_mean(res, prior_tol = 1e-09)

susie_get_posterior_sd(res, prior_tol = 1e-09)

susie_get_niter(res)

susie_get_prior_variance(res)

susie_get_residual_variance(res)

susie_get_lfsr(res)

susie_get_posterior_samples(susie_fit, num_samples)

susie_get_cs(
  res,
  X = NULL,
  Xcorr = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  dedup = TRUE,
  squared = FALSE,
  check_symmetric = TRUE,
  n_purity = 100,
  use_rfast
)

susie_get_pip(res, prune_by_cs = FALSE, prior_tol = 1e-09)

Arguments

res

A susie fit, typically an output from susie or one of its variants. For susie_get_pip and susie_get_cs, this may instead be the posterior inclusion probability matrix, alpha.

last_only

If last_only = FALSE, return the ELBO from all iterations; otherwise return the ELBO from the last iteration only.

warning_tol

Warn if ELBO is decreasing by this tolerance level.

prior_tol

Filter out effects having estimated prior variance smaller than this threshold.

susie_fit

A susie fit, an output from susie.

num_samples

The number of draws from the posterior distribution.

X

n by p matrix of values of the p variables (covariates) in n samples. When provided, correlation between variables will be computed and used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.

Xcorr

p by p matrix of correlations between variables (covariates). When provided, it will be used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.

coverage

A number between 0 and 1 specifying desired coverage of each CS.

min_abs_corr

A "purity" threshold for the CS. Any CS that contains a pair of variables with correlation less than this threshold will be filtered out and not reported.

dedup

If dedup = TRUE, remove duplicate CSs.

squared

If squared = TRUE, report min, mean and median of squared correlation instead of the absolute correlation.

check_symmetric

If check_symmetric = TRUE, perform a check for symmetry of matrix Xcorr when Xcorr is provided (not NULL).

n_purity

The maximum number of credible set (CS) variables used in calculating the correlation (“purity”) statistics. When the number of variables included in the CS is greater than this number, the CS variables are randomly subsampled.

use_rfast

Use the Rfast package for the purity calculations. By default use_rfast = TRUE if the Rfast package is installed.

prune_by_cs

Whether or not to ignore single effects not in a reported CS when calculating PIP.

Value

susie_get_objective returns the evidence lower bound (ELBO) achieved by the fitted susie model and, optionally, at each iteration of the IBSS fitting procedure.

susie_get_residual_variance returns the (estimated or fixed) residual variance parameter.

susie_get_prior_variance returns the (estimated or fixed) prior variance parameters.

susie_get_posterior_mean returns the posterior mean for the regression coefficients of the fitted susie model.

susie_get_posterior_sd returns the posterior standard deviation for coefficients of the fitted susie model.

susie_get_niter returns the number of model fitting iterations performed.

susie_get_pip returns a vector containing the posterior inclusion probabilities (PIPs) for all variables.

susie_get_lfsr returns a vector containing the average lfsr across variables for each single-effect, weighted by the posterior inclusion probability (alpha).

susie_get_posterior_samples returns a list containing the effect sizes samples and causal status with two components: b, an num_variables x num_samples matrix of effect sizes; gamma, an num_variables x num_samples matrix of causal status random draws.

susie_get_cs returns credible sets (CSs) from a susie fit, as well as summaries of correlation among the variables included in each CS. If desired, one can filter out CSs that do not meet a specified “purity” threshold; to do this, either X or Xcorr must be supplied. It returns a list with the following elements:

cs

A list in which each list element is a vector containing the indices of the variables in the CS.

coverage

The nominal coverage specified for each CS.

purity

If X or Xcorr iis provided), the purity of each CS.

cs_index

If X or Xcorr is provided) the index (number between 1 and L) of each reported CS in the supplied susie fit.

Examples

set.seed(1)
n <- 1000
p <- 1000
beta <- rep(0, p)
beta[1:4] <- 1
X <- matrix(rnorm(n * p), nrow = n, ncol = p)
X <- scale(X, center = TRUE, scale = TRUE)
y <- drop(X %*% beta + rnorm(n))
s <- susie(X, y, L = 10)
susie_get_objective(s)
#> [1] -1455
susie_get_objective(s, last_only = FALSE)
#> [1] -1927 -1478 -1455 -1455
susie_get_residual_variance(s)
#> [1] 0.9899
susie_get_prior_variance(s)
#>  [1] 1.1327 1.0495 1.0094 0.9784 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
susie_get_posterior_mean(s)
#>    [1]   1.064e+00   1.024e+00   1.004e+00   9.887e-01 -8.453e-217 -1.417e-213
#>    [7]  1.202e-214  4.562e-216  1.734e-216 -4.658e-217  2.154e-216  5.615e-216
#>   [13]  1.120e-216  2.863e-216 -1.154e-216  1.523e-218  1.217e-215  2.958e-215
#>   [19]  2.874e-216 -3.406e-216 -4.755e-215 -1.674e-215  2.575e-215 -4.179e-216
#>   [25]  5.354e-216 -1.398e-216  9.686e-217  1.577e-217  5.922e-217 -1.358e-216
#>   [31] -2.469e-217  8.746e-214  8.298e-217 -6.323e-217 -4.609e-217  1.160e-212
#>   [37]  3.595e-217 -3.884e-216 -1.544e-216  4.806e-216  5.560e-217  9.513e-218
#>   [43] -9.188e-217 -9.134e-216 -5.319e-216  1.608e-213 -6.428e-215  2.395e-216
#>   [49] -2.447e-217  3.164e-215 -1.420e-215 -2.331e-217  1.873e-217 -1.019e-216
#>   [55]  1.511e-216  4.938e-216  7.497e-217 -5.946e-217 -6.349e-216  1.211e-215
#>   [61]  7.046e-217 -2.809e-215  3.566e-217 -6.750e-217 -2.882e-217  5.030e-217
#>   [67]  1.203e-215  1.894e-216 -1.304e-212  5.339e-217  8.686e-216  1.800e-214
#>   [73]  2.854e-217 -5.238e-215 -3.621e-216 -4.333e-217  1.144e-216  1.609e-215
#>   [79]  4.164e-214 -2.362e-215 -7.198e-217  6.355e-217  4.804e-216 -1.581e-217
#>   [85]  3.803e-216 -3.296e-216 -3.715e-218  6.677e-217 -3.206e-216  2.746e-216
#>   [91] -4.360e-216 -5.909e-215  6.769e-217  1.817e-216 -4.795e-217  8.462e-217
#>   [97] -7.123e-217  1.489e-215 -8.621e-216 -1.364e-216  1.712e-217  8.757e-216
#>  [103] -1.907e-214 -1.912e-216  3.787e-216  3.310e-214 -1.766e-216 -5.336e-216
#>  [109]  6.925e-217  2.333e-216  4.734e-217  2.230e-215 -6.674e-216  3.678e-218
#>  [115]  5.456e-216  1.432e-216 -2.774e-217  1.210e-216 -9.883e-217  7.836e-216
#>  [121]  4.984e-217 -1.352e-216  8.744e-217 -1.519e-215  9.001e-217 -7.814e-217
#>  [127] -1.199e-215  4.466e-216  6.151e-217  3.082e-215  2.764e-215  5.335e-217
#>  [133] -4.453e-217 -3.049e-214  3.333e-217  4.060e-217  6.812e-218 -3.099e-216
#>  [139]  5.098e-216  5.813e-217 -5.065e-217  1.005e-216  3.098e-217 -5.377e-217
#>  [145] -1.028e-216 -1.439e-215  1.519e-214  1.169e-216  4.909e-216 -5.611e-216
#>  [151]  3.416e-217  1.954e-216  2.989e-216  5.508e-217  2.091e-217 -2.408e-215
#>  [157] -1.364e-216 -2.112e-216 -1.136e-215 -1.626e-216 -9.848e-217 -9.760e-215
#>  [163] -1.407e-214  4.288e-217 -3.571e-216  7.880e-215 -6.005e-217  1.886e-217
#>  [169]  4.400e-216  1.916e-214  2.074e-216  7.297e-217 -3.008e-216 -2.471e-216
#>  [175] -5.289e-216 -7.912e-217 -2.159e-216  2.344e-215 -3.252e-216  7.733e-215
#>  [181]  1.057e-217 -8.224e-216 -1.835e-216  2.387e-216  1.495e-215  2.548e-217
#>  [187] -3.764e-215 -1.204e-217 -1.796e-216 -2.285e-215 -4.365e-217 -1.154e-216
#>  [193]  7.928e-216  2.673e-217  3.477e-217 -1.767e-216 -1.097e-216 -2.950e-216
#>  [199] -1.046e-216  1.260e-216  5.504e-216 -2.288e-214  1.018e-216 -4.558e-215
#>  [205]  3.710e-215 -7.384e-217  3.732e-216  1.205e-215 -1.101e-217 -2.296e-215
#>  [211] -8.247e-217  3.226e-218 -1.082e-216 -1.359e-213 -4.782e-216 -1.532e-217
#>  [217] -1.732e-214 -5.208e-216  1.604e-215 -1.357e-217 -8.577e-217 -8.091e-217
#>  [223] -8.123e-217  7.593e-217 -6.029e-216  1.187e-216  2.929e-216 -1.593e-215
#>  [229] -6.386e-216 -4.185e-213  2.569e-216 -1.212e-216  2.458e-216  8.303e-216
#>  [235]  1.863e-215 -3.387e-216  1.661e-217 -9.643e-216  5.009e-216 -1.763e-215
#>  [241]  6.914e-217 -9.057e-218  8.903e-216  4.777e-216  7.994e-217  3.788e-215
#>  [247]  4.257e-217 -1.322e-216  5.918e-218  1.741e-216 -5.331e-217  1.516e-217
#>  [253] -1.208e-216 -1.322e-216 -4.289e-217  2.335e-216  4.526e-217 -7.226e-217
#>  [259]  4.874e-217  4.105e-217 -1.126e-216 -4.643e-217 -5.811e-217 -6.008e-216
#>  [265]  2.123e-215  2.595e-217 -5.697e-216 -1.808e-216  3.316e-216  2.289e-216
#>  [271] -1.477e-216 -1.635e-217 -1.439e-212 -2.938e-217  2.575e-213 -1.465e-216
#>  [277]  1.349e-216  6.645e-219 -3.520e-216  2.706e-216 -1.906e-215 -8.882e-216
#>  [283]  8.199e-217 -3.245e-216  1.364e-218 -2.396e-218 -1.865e-216  2.517e-217
#>  [289] -8.451e-218  8.603e-217 -1.163e-217  1.604e-214 -4.195e-217  2.919e-217
#>  [295] -2.790e-216  4.070e-216 -4.880e-215 -8.956e-216  5.085e-217  3.925e-216
#>  [301] -3.890e-218 -9.459e-217  9.715e-213 -1.212e-216  7.154e-217 -3.993e-216
#>  [307]  7.714e-216 -2.997e-216 -2.601e-215  2.139e-216  7.738e-217 -6.064e-219
#>  [313]  1.043e-216  1.165e-216  4.820e-216  1.805e-215 -4.075e-216  3.446e-216
#>  [319]  3.054e-212 -1.346e-216  4.549e-217 -1.007e-215  9.171e-218  6.648e-216
#>  [325] -6.458e-217  3.211e-215 -2.143e-216  6.248e-217 -1.096e-215 -1.588e-216
#>  [331]  2.121e-217  1.634e-215  6.041e-216  6.388e-218  1.308e-217  6.461e-215
#>  [337] -8.267e-216  4.564e-216  1.617e-217 -1.133e-216  5.229e-216 -2.475e-217
#>  [343]  3.229e-216 -7.205e-216  8.456e-217  1.557e-214  4.271e-216 -1.606e-216
#>  [349]  1.618e-217  8.312e-216  3.268e-216 -3.658e-216  6.369e-216 -4.042e-215
#>  [355]  1.208e-215 -6.175e-217 -9.996e-215 -1.111e-216 -7.325e-214 -1.243e-217
#>  [361] -3.296e-215  1.706e-215  2.573e-218 -2.298e-214 -9.716e-215 -2.964e-216
#>  [367] -1.209e-216  1.435e-216  3.267e-215  5.341e-217 -7.211e-216  1.982e-217
#>  [373]  8.753e-217 -4.585e-215  1.292e-216  2.719e-218 -5.300e-216  5.186e-216
#>  [379] -2.882e-216  8.720e-216 -5.496e-217 -1.586e-216  8.673e-214  3.471e-217
#>  [385] -1.459e-214  2.768e-217 -1.154e-216 -2.106e-216 -7.965e-218  4.591e-217
#>  [391]  1.158e-214  1.434e-216 -7.786e-217 -6.112e-217  7.357e-217 -5.243e-216
#>  [397] -1.568e-213  5.712e-217  7.237e-218  1.539e-215  1.578e-216 -2.141e-216
#>  [403]  2.737e-216 -1.428e-215  3.914e-217 -1.835e-216 -3.290e-216 -5.781e-217
#>  [409] -6.155e-215 -3.402e-216 -3.677e-216  1.315e-216  4.345e-216  4.522e-217
#>  [415] -8.248e-215 -1.721e-216 -3.842e-216  1.385e-215 -1.800e-216 -8.137e-214
#>  [421]  7.291e-216  2.836e-216  2.884e-216 -4.488e-217 -3.741e-216 -6.680e-216
#>  [427]  5.024e-217 -1.756e-215  4.303e-216 -1.938e-216 -3.806e-217  3.832e-217
#>  [433]  1.657e-214  1.956e-216  2.881e-216 -1.846e-216  2.415e-216  1.003e-216
#>  [439]  5.749e-217 -1.897e-216  2.137e-216  4.608e-217 -1.843e-216 -3.532e-216
#>  [445] -1.096e-215 -4.185e-217 -2.492e-217 -9.836e-216  2.717e-217  1.109e-216
#>  [451] -4.324e-214 -2.981e-216  4.333e-216  1.054e-216  9.365e-216  1.832e-216
#>  [457]  4.113e-215 -3.087e-216  3.701e-215 -3.374e-216  1.060e-215  7.330e-216
#>  [463] -2.977e-216  1.623e-216 -2.116e-217  7.261e-217 -4.886e-217  5.659e-216
#>  [469]  1.702e-215  2.197e-215 -7.433e-216  1.298e-216 -2.672e-216  3.721e-218
#>  [475]  3.770e-216 -1.537e-216 -3.987e-216  1.567e-215 -9.148e-217  1.836e-216
#>  [481] -1.165e-215 -4.961e-215  1.089e-215 -6.668e-217  2.002e-217 -1.650e-213
#>  [487] -1.703e-216 -8.431e-217 -6.197e-216 -3.554e-215 -8.758e-217 -6.050e-217
#>  [493]  1.407e-217 -5.779e-216  5.222e-218 -1.207e-216  1.236e-216  9.301e-217
#>  [499] -7.571e-217  1.006e-215  2.853e-217 -4.660e-215  2.067e-215  3.180e-217
#>  [505]  4.637e-217 -8.935e-217 -1.909e-217  2.097e-216 -4.629e-216  3.938e-216
#>  [511] -1.674e-215 -5.241e-215  2.084e-216  4.789e-218 -1.149e-214 -1.860e-215
#>  [517]  3.041e-216  1.632e-216 -1.175e-214 -7.497e-217  5.921e-218 -4.512e-216
#>  [523] -2.941e-218  7.982e-218  9.944e-217  1.934e-214  4.159e-217 -6.898e-217
#>  [529]  3.399e-215 -2.642e-216  7.784e-216  4.275e-216 -4.110e-214 -1.988e-216
#>  [535]  5.945e-217  1.339e-216 -4.732e-216  1.076e-216 -4.003e-216 -5.017e-217
#>  [541]  1.606e-218  2.404e-216 -6.845e-219 -5.627e-216  2.536e-216 -1.177e-216
#>  [547] -1.031e-216  5.228e-216  1.364e-213  1.781e-216  1.265e-216 -5.776e-217
#>  [553] -9.490e-217 -1.345e-214 -2.059e-211  8.442e-217  1.581e-217  3.423e-216
#>  [559]  1.415e-216  4.247e-216  8.564e-217 -4.449e-216  4.526e-217  8.187e-217
#>  [565]  3.359e-214 -4.655e-217 -4.594e-216 -4.403e-215  6.124e-216  3.392e-217
#>  [571] -2.741e-215 -6.880e-216 -6.873e-217 -7.492e-217  3.127e-216  3.580e-217
#>  [577]  4.040e-216  2.167e-216  2.651e-216 -9.424e-218  2.645e-216  4.085e-216
#>  [583] -2.097e-216  4.239e-216 -2.250e-216 -1.455e-215 -2.185e-217  1.107e-216
#>  [589]  2.802e-217  6.746e-216  6.531e-216 -4.163e-216  4.142e-216  4.830e-214
#>  [595] -1.514e-216 -3.472e-216  2.116e-216  1.162e-217  1.416e-216  2.052e-215
#>  [601] -6.663e-218  1.022e-216  2.033e-215  4.220e-217  7.624e-216  1.116e-216
#>  [607]  1.296e-217 -1.332e-215  3.170e-217  1.671e-215  3.154e-217  3.557e-211
#>  [613] -5.450e-217 -6.054e-216 -6.967e-218 -1.568e-216 -6.441e-217  2.594e-215
#>  [619] -1.154e-215  4.526e-217  9.337e-216  9.793e-217 -4.077e-217 -1.222e-214
#>  [625] -2.758e-215 -1.959e-215 -1.571e-216  6.620e-216 -1.548e-216 -1.437e-217
#>  [631] -1.592e-215  4.920e-217 -4.120e-216 -6.558e-217  1.317e-216 -8.103e-215
#>  [637]  4.865e-216 -1.229e-216  3.074e-217 -9.863e-216  5.901e-215  7.799e-218
#>  [643] -4.146e-213 -6.453e-216  2.409e-217 -5.737e-217 -2.461e-217 -1.446e-216
#>  [649]  2.617e-216 -1.818e-216  5.528e-217  4.256e-217  2.483e-216 -1.258e-217
#>  [655]  6.052e-216 -9.545e-217 -9.766e-217  6.115e-216 -2.390e-217  6.728e-217
#>  [661]  9.469e-216  2.350e-217 -7.240e-215  1.220e-216  1.427e-216  6.120e-217
#>  [667] -3.009e-216 -6.908e-217 -9.117e-217  1.276e-216  1.379e-215 -5.335e-218
#>  [673]  1.853e-216 -3.208e-217  4.038e-217  6.814e-217  1.504e-217 -1.191e-216
#>  [679] -1.132e-217  7.253e-216  6.948e-217  7.273e-216 -3.307e-216 -8.055e-217
#>  [685] -2.168e-216 -1.693e-213 -1.422e-216 -2.858e-218  1.531e-213  1.590e-216
#>  [691]  3.052e-215  3.558e-216  2.195e-217  1.192e-216 -4.440e-217 -8.613e-217
#>  [697] -4.417e-219  1.758e-217  2.842e-216 -9.495e-217 -1.483e-215 -1.284e-215
#>  [703] -1.954e-216  1.048e-216  1.400e-216  1.027e-215  9.451e-216  8.970e-217
#>  [709] -3.023e-216 -3.224e-215  6.141e-215  1.178e-214  8.439e-217 -3.838e-216
#>  [715]  1.762e-216 -4.924e-215  4.115e-216 -1.508e-216 -8.554e-217 -7.506e-216
#>  [721]  5.486e-217  1.231e-216 -6.691e-217  5.906e-215 -2.106e-216  2.607e-217
#>  [727] -8.027e-215 -9.365e-214  7.215e-219  1.641e-216 -1.806e-216  2.753e-213
#>  [733] -1.465e-216 -6.659e-216 -7.030e-216 -7.846e-217 -3.815e-216  7.517e-216
#>  [739] -2.883e-216  1.415e-216 -6.471e-216 -5.246e-217  4.287e-217  3.423e-218
#>  [745]  1.364e-215 -2.818e-215 -1.073e-216 -4.544e-216 -1.911e-215  8.147e-217
#>  [751]  3.264e-216 -6.613e-216  5.652e-217 -5.789e-217 -2.998e-217 -4.056e-216
#>  [757] -3.205e-215  2.776e-214 -6.441e-218  2.078e-215  2.183e-217  1.193e-216
#>  [763] -6.354e-215  2.217e-216 -1.605e-215  2.377e-214  1.572e-216  5.474e-216
#>  [769] -1.412e-216  5.224e-217  1.144e-215  3.239e-216 -1.686e-216  1.325e-213
#>  [775] -2.971e-217  1.139e-218  1.615e-216 -2.817e-216 -3.192e-216  2.332e-216
#>  [781]  6.889e-217  1.205e-216  3.018e-216  1.632e-216 -2.164e-215 -3.444e-217
#>  [787]  2.417e-216  5.417e-217 -1.193e-215  1.159e-217 -2.610e-216  4.061e-216
#>  [793] -3.732e-214  8.425e-218  1.094e-215 -4.507e-217 -7.880e-216 -3.167e-215
#>  [799] -7.731e-215 -4.619e-216  5.273e-217  8.358e-215 -1.211e-216  2.179e-216
#>  [805] -3.601e-217 -9.033e-217  8.012e-217  4.438e-217 -1.061e-215  3.954e-215
#>  [811]  1.134e-216  1.203e-215  6.196e-216  1.628e-216 -3.270e-216 -2.701e-217
#>  [817] -2.189e-216  1.882e-214  4.402e-217 -3.261e-215 -8.172e-216 -1.043e-215
#>  [823] -3.726e-217 -4.258e-216 -3.902e-216 -3.642e-216 -1.759e-217 -5.189e-217
#>  [829]  1.246e-217  5.659e-217  5.509e-216 -6.699e-216  2.867e-215  1.767e-216
#>  [835]  1.683e-217  3.721e-217 -9.504e-217 -1.043e-216  5.945e-216  1.680e-216
#>  [841] -2.593e-216  2.451e-216  2.325e-216  1.616e-216  6.039e-216  1.946e-216
#>  [847] -2.621e-217 -2.446e-216  2.469e-216  1.498e-217  4.117e-217  3.280e-215
#>  [853]  1.272e-217  2.449e-216  5.139e-217 -2.443e-215  1.789e-217  6.552e-217
#>  [859]  2.466e-216 -8.745e-216 -3.366e-216 -2.215e-217  3.187e-215 -2.272e-217
#>  [865]  2.156e-216  5.080e-216 -1.274e-215  5.332e-217 -1.838e-216  5.548e-216
#>  [871] -3.494e-217 -2.838e-216 -1.103e-215 -2.322e-216  1.421e-216  7.950e-217
#>  [877]  8.453e-216  9.113e-215  3.034e-217 -1.781e-217  5.151e-216 -1.794e-216
#>  [883]  1.182e-215  5.064e-217 -4.209e-216 -1.511e-217 -2.603e-216 -1.594e-217
#>  [889]  4.202e-217  8.254e-217 -1.811e-215  7.633e-217 -1.540e-216 -2.192e-217
#>  [895] -1.164e-217  2.039e-216  6.115e-215 -6.188e-215 -3.952e-216 -8.386e-217
#>  [901]  1.072e-214  1.342e-216  1.288e-215  5.225e-216 -1.348e-215 -7.381e-215
#>  [907] -3.411e-216  6.117e-216 -1.148e-216  1.429e-216 -4.739e-217 -2.171e-216
#>  [913] -1.742e-216  1.019e-215  7.904e-213  2.302e-218 -4.385e-217 -1.062e-217
#>  [919] -1.014e-216 -2.021e-216 -2.222e-216 -1.554e-216  8.252e-216  3.375e-217
#>  [925] -3.747e-217  1.713e-217 -1.046e-217 -1.748e-217 -4.350e-217 -3.280e-215
#>  [931]  2.439e-216 -4.339e-216  6.351e-216 -2.825e-213  1.634e-216  1.070e-217
#>  [937] -2.002e-216 -1.814e-216  2.827e-216 -5.622e-216 -3.668e-217  5.851e-216
#>  [943] -2.138e-217  1.392e-216 -2.552e-214 -4.220e-216  3.836e-215  7.467e-217
#>  [949]  3.764e-218  2.732e-216 -5.217e-217 -4.602e-216  8.365e-217 -1.252e-216
#>  [955]  1.053e-216 -2.538e-217 -8.972e-217  2.475e-218 -3.767e-215 -5.090e-216
#>  [961]  1.677e-215  2.307e-216  2.328e-216 -6.006e-216 -4.529e-215  1.421e-217
#>  [967] -6.985e-213  1.628e-217 -3.268e-217 -3.994e-217  5.126e-217 -4.329e-217
#>  [973]  2.387e-217 -3.940e-215 -7.916e-216 -1.818e-216 -8.943e-214  1.075e-217
#>  [979]  1.240e-216  1.811e-215  8.493e-217 -8.261e-216  4.701e-217 -1.601e-215
#>  [985] -7.171e-217  1.894e-217 -1.181e-216 -6.571e-217 -2.003e-216 -7.260e-216
#>  [991] -1.257e-216  1.893e-216 -3.397e-214 -1.739e-215  7.822e-218  9.531e-214
#>  [997] -7.283e-217  2.477e-216 -1.300e-216  1.246e-214
susie_get_posterior_sd(s)
#>    [1]  3.146e-02  3.146e-02  3.146e-02  3.146e-02 2.472e-109 1.286e-107
#>    [7] 3.399e-108 5.509e-109 3.309e-109 2.150e-109 3.683e-109 6.185e-109
#>   [13] 2.731e-109 4.272e-109 2.764e-109 1.961e-109 9.552e-109 1.570e-108
#>   [19] 4.281e-109 4.691e-109 2.043e-108 1.142e-108 1.453e-108 5.248e-109
#>   [25] 6.022e-109 2.995e-109 2.587e-109 1.985e-109 2.248e-109 2.957e-109
#>   [31] 2.019e-109 9.934e-108 2.458e-109 2.282e-109 2.146e-109 3.915e-107
#>   [37] 2.079e-109 5.041e-109 3.133e-109 5.671e-109 2.219e-109 1.970e-109
#>   [43] 2.540e-109 8.128e-109 6.001e-109 1.375e-107 2.412e-108 3.889e-109
#>   [49] 2.018e-109 1.630e-108 1.042e-108 2.013e-109 1.995e-109 2.635e-109
#>   [55] 3.102e-109 5.757e-109 2.385e-109 2.250e-109 6.625e-109 9.525e-109
#>   [61] 2.344e-109 1.525e-108 2.077e-109 2.318e-109 2.039e-109 2.178e-109
#>   [67] 9.490e-109 3.453e-109 4.164e-107 2.201e-109 7.901e-109 4.237e-108
#>   [73] 2.037e-109 2.155e-108 4.850e-109 2.127e-109 2.755e-109 1.117e-108
#>   [79] 6.669e-108 1.385e-108 2.358e-109 2.285e-109 5.670e-109 1.985e-109
#>   [85] 4.982e-109 4.609e-109 1.962e-109 2.312e-109 4.540e-109 4.178e-109
#>   [91] 5.373e-109 2.303e-108 2.320e-109 3.384e-109 2.160e-109 2.473e-109
#>   [97] 2.351e-109 1.070e-108 7.868e-109 2.963e-109 1.989e-109 7.937e-109
#>  [103] 4.371e-108 3.469e-109 4.971e-109 5.893e-108 3.338e-109 6.011e-109
#>  [109] 2.334e-109 3.836e-109 2.156e-109 1.341e-108 6.814e-109 1.962e-109
#>  [115] 6.086e-109 3.028e-109 2.033e-109 2.818e-109 2.606e-109 7.456e-109
#>  [121] 2.174e-109 2.952e-109 2.499e-109 1.082e-108 2.523e-109 2.413e-109
#>  [127] 9.471e-109 5.445e-109 2.267e-109 1.606e-108 1.512e-108 2.201e-109
#>  [133] 2.135e-109 5.637e-108 2.063e-109 2.108e-109 1.965e-109 4.457e-109
#>  [139] 5.860e-109 2.239e-109 2.180e-109 2.622e-109 2.050e-109 2.204e-109
#>  [145] 2.643e-109 1.049e-108 3.862e-108 2.778e-109 5.738e-109 6.183e-109
#>  [151] 2.068e-109 3.507e-109 4.372e-109 2.215e-109 2.003e-109 1.400e-108
#>  [157] 2.964e-109 3.646e-109 9.186e-109 3.208e-109 2.603e-109 3.034e-108
#>  [163] 3.705e-108 2.124e-109 4.814e-109 2.698e-108 2.255e-109 1.995e-109
#>  [169] 5.400e-109 4.383e-108 3.613e-109 2.367e-109 4.386e-109 3.953e-109
#>  [175] 6.200e-109 2.422e-109 3.687e-109 1.379e-108 4.575e-109 2.670e-108
#>  [181] 1.972e-109 7.662e-109 3.400e-109 3.882e-109 1.072e-108 2.022e-109
#>  [187] 1.795e-108 1.975e-109 3.364e-109 1.359e-108 2.129e-109 2.763e-109
#>  [193] 7.505e-109 2.028e-109 2.072e-109 3.338e-109 2.709e-109 4.341e-109
#>  [199] 2.660e-109 2.865e-109 6.116e-109 4.826e-108 2.634e-109 1.996e-108
#>  [205] 1.780e-108 2.374e-109 4.931e-109 9.496e-109 1.973e-109 1.363e-108
#>  [211] 2.453e-109 1.962e-109 2.695e-109 1.257e-107 5.655e-109 1.984e-109
#>  [217] 4.149e-108 5.930e-109 1.115e-108 1.979e-109 2.483e-109 2.438e-109
#>  [223] 2.441e-109 2.393e-109 6.436e-109 2.795e-109 4.324e-109 1.111e-108
#>  [229] 6.647e-109 2.285e-107 4.034e-109 2.819e-109 3.941e-109 7.703e-109
#>  [235] 1.213e-108 4.677e-109 1.987e-109 8.379e-109 5.803e-109 1.176e-108
#>  [241] 2.333e-109 1.969e-109 8.011e-109 5.652e-109 2.430e-109 1.801e-108
#>  [247] 2.122e-109 2.924e-109 1.964e-109 3.314e-109 2.201e-109 1.983e-109
#>  [253] 2.816e-109 2.924e-109 2.124e-109 3.838e-109 2.141e-109 2.360e-109
#>  [259] 2.166e-109 2.111e-109 2.737e-109 2.149e-109 2.239e-109 6.424e-109
#>  [265] 1.305e-108 2.024e-109 6.235e-109 3.376e-109 4.623e-109 3.799e-109
#>  [271] 3.070e-109 1.987e-109 4.385e-107 2.041e-109 1.766e-107 3.059e-109
#>  [277] 2.950e-109 1.961e-109 4.776e-109 4.146e-109 1.228e-108 8.000e-109
#>  [283] 2.448e-109 4.570e-109 1.961e-109 1.961e-109 3.427e-109 2.021e-109
#>  [289] 1.968e-109 2.486e-109 1.974e-109 3.979e-108 2.117e-109 2.040e-109
#>  [295] 4.214e-109 5.172e-109 2.072e-108 8.038e-109 2.182e-109 5.070e-109
#>  [301] 1.962e-109 2.566e-109 3.566e-107 2.819e-109 2.354e-109 5.118e-109
#>  [307] 7.391e-109 4.378e-109 1.461e-108 3.669e-109 2.406e-109 1.961e-109
#>  [313] 2.658e-109 2.775e-109 5.680e-109 1.191e-108 5.175e-109 4.721e-109
#>  [319] 6.510e-107 2.947e-109 2.142e-109 8.586e-109 1.969e-109 6.799e-109
#>  [325] 2.293e-109 1.643e-108 3.673e-109 2.275e-109 9.007e-109 3.174e-109
#>  [331] 2.004e-109 1.127e-108 6.443e-109 1.965e-109 1.977e-109 2.419e-108
#>  [337] 7.684e-109 5.510e-109 1.986e-109 2.743e-109 5.944e-109 2.019e-109
#>  [343] 4.558e-109 7.112e-109 2.472e-109 3.915e-108 5.312e-109 3.191e-109
#>  [349] 1.986e-109 7.707e-109 4.587e-109 4.877e-109 6.637e-109 1.867e-108
#>  [355] 9.510e-109 2.269e-109 3.074e-108 2.722e-109 9.034e-108 1.976e-109
#>  [361] 1.667e-108 1.154e-108 1.961e-109 4.837e-108 3.026e-108 4.352e-109
#>  [367] 2.816e-109 3.030e-109 1.659e-108 2.202e-109 7.116e-109 1.999e-109
#>  [373] 2.500e-109 2.002e-108 2.895e-109 1.962e-109 5.989e-109 5.917e-109
#>  [379] 4.287e-109 7.918e-109 2.214e-109 3.172e-109 9.890e-108 2.071e-109
#>  [385] 3.779e-108 2.033e-109 2.764e-109 3.641e-109 1.967e-109 2.145e-109
#>  [391] 3.332e-108 3.029e-109 2.411e-109 2.264e-109 2.372e-109 5.952e-109
#>  [397] 1.357e-107 2.231e-109 1.966e-109 1.090e-108 3.164e-109 3.671e-109
#>  [403] 4.171e-109 1.045e-108 2.099e-109 3.400e-109 4.604e-109 2.237e-109
#>  [409] 2.355e-108 4.689e-109 4.892e-109 2.917e-109 5.362e-109 2.140e-109
#>  [415] 2.767e-108 3.297e-109 5.011e-109 1.027e-108 3.368e-109 9.558e-108
#>  [421] 7.160e-109 4.251e-109 4.288e-109 2.138e-109 4.938e-109 6.817e-109
#>  [427] 2.177e-109 1.173e-108 5.334e-109 3.492e-109 2.092e-109 2.094e-109
#>  [433] 4.050e-108 3.508e-109 4.287e-109 3.410e-109 3.905e-109 2.620e-109
#>  [439] 2.234e-109 3.456e-109 3.668e-109 2.146e-109 3.407e-109 4.785e-109
#>  [445] 9.003e-109 2.117e-109 2.020e-109 8.473e-109 2.030e-109 2.721e-109
#>  [451] 6.806e-108 4.365e-109 5.354e-109 2.668e-109 8.242e-109 3.397e-109
#>  [457] 1.885e-108 4.448e-109 1.778e-108 4.667e-109 8.835e-109 7.182e-109
#>  [463] 4.362e-109 3.206e-109 2.004e-109 2.363e-109 2.167e-109 6.212e-109
#>  [469] 1.153e-108 1.330e-108 7.238e-109 2.901e-109 4.118e-109 1.962e-109
#>  [475] 4.959e-109 3.126e-109 5.114e-109 1.101e-108 2.537e-109 3.401e-109
#>  [481] 9.321e-109 2.091e-108 8.973e-109 2.311e-109 1.999e-109 1.394e-107
#>  [487] 3.280e-109 2.470e-109 6.536e-109 1.739e-108 2.500e-109 2.259e-109
#>  [493] 1.980e-109 6.285e-109 1.963e-109 2.814e-109 2.842e-109 2.551e-109
#>  [499] 2.391e-109 8.581e-109 2.037e-109 2.020e-108 1.286e-108 2.054e-109
#>  [505] 2.148e-109 2.517e-109 1.996e-109 3.633e-109 5.554e-109 5.079e-109
#>  [511] 1.142e-108 2.155e-108 3.621e-109 1.963e-109 3.317e-108 1.212e-108
#>  [517] 4.412e-109 3.215e-109 3.358e-108 2.385e-109 1.964e-109 5.476e-109
#>  [523] 1.962e-109 1.967e-109 2.612e-109 4.404e-108 2.115e-109 2.331e-109
#>  [529] 1.696e-108 4.094e-109 7.429e-109 5.314e-109 6.622e-108 3.537e-109
#>  [535] 2.250e-109 2.940e-109 5.622e-109 2.689e-109 5.125e-109 2.177e-109
#>  [541] 1.961e-109 3.897e-109 1.961e-109 6.192e-109 4.006e-109 2.785e-109
#>  [547] 2.646e-109 5.943e-109 1.260e-107 3.351e-109 2.869e-109 2.236e-109
#>  [553] 2.569e-109 3.615e-108 1.764e-106 2.471e-109 1.985e-109 4.704e-109
#>  [559] 3.011e-109 5.295e-109 2.482e-109 5.433e-109 2.140e-109 2.447e-109
#>  [565] 5.940e-108 2.150e-109 5.530e-109 1.958e-108 6.493e-109 2.067e-109
#>  [571] 1.505e-108 6.931e-109 2.329e-109 2.384e-109 4.479e-109 2.078e-109
#>  [577] 5.151e-109 3.694e-109 4.101e-109 1.969e-109 4.096e-109 5.183e-109
#>  [583] 3.633e-109 5.290e-109 3.766e-109 1.056e-108 2.006e-109 2.719e-109
#>  [589] 2.034e-109 6.855e-109 6.731e-109 5.237e-109 5.222e-109 7.223e-108
#>  [595] 3.105e-109 4.741e-109 3.650e-109 1.974e-109 3.013e-109 1.280e-108
#>  [601] 1.965e-109 2.638e-109 1.274e-108 2.119e-109 7.342e-109 2.727e-109
#>  [607] 1.977e-109 1.005e-108 2.054e-109 1.141e-108 2.053e-109 2.344e-106
#>  [613] 2.210e-109 6.451e-109 1.966e-109 3.155e-109 2.292e-109 1.459e-108
#>  [619] 9.268e-109 2.141e-109 8.229e-109 2.597e-109 2.110e-109 3.431e-108
#>  [625] 1.510e-108 1.248e-108 3.158e-109 6.782e-109 3.137e-109 1.981e-109
#>  [631] 1.111e-108 2.169e-109 5.207e-109 2.302e-109 2.919e-109 2.740e-108
#>  [637] 5.710e-109 2.836e-109 2.049e-109 8.486e-109 2.301e-108 1.967e-109
#>  [643] 2.274e-107 6.686e-109 2.016e-109 2.233e-109 2.018e-109 3.041e-109
#>  [649] 4.073e-109 3.385e-109 2.216e-109 2.122e-109 3.963e-109 1.976e-109
#>  [655] 6.450e-109 2.574e-109 2.595e-109 6.487e-109 2.015e-109 2.317e-109
#>  [661] 8.294e-109 2.013e-109 2.575e-108 2.827e-109 3.023e-109 2.265e-109
#>  [667] 4.387e-109 2.332e-109 2.534e-109 2.880e-109 1.024e-108 1.964e-109
#>  [673] 3.416e-109 2.056e-109 2.107e-109 2.324e-109 1.983e-109 2.799e-109
#>  [679] 1.973e-109 7.139e-109 2.336e-109 7.151e-109 4.617e-109 2.435e-109
#>  [685] 3.695e-109 1.413e-107 3.019e-109 1.962e-109 1.340e-107 3.175e-109
#>  [691] 1.598e-108 4.804e-109 2.007e-109 2.800e-109 2.134e-109 2.487e-109
#>  [697] 1.961e-109 1.991e-109 4.255e-109 2.569e-109 1.067e-108 9.843e-109
#>  [703] 3.507e-109 2.663e-109 2.997e-109 8.680e-109 8.285e-109 2.520e-109
#>  [709] 4.398e-109 1.647e-108 2.352e-108 3.363e-108 2.470e-109 5.008e-109
#>  [715] 3.333e-109 2.083e-108 5.203e-109 3.099e-109 2.481e-109 7.278e-109
#>  [721] 2.213e-109 2.838e-109 2.313e-109 2.302e-108 3.641e-109 2.025e-109
#>  [727] 2.726e-108 1.030e-107 1.961e-109 3.223e-109 3.374e-109 1.830e-107
#>  [733] 3.059e-109 6.805e-109 7.015e-109 2.416e-109 4.991e-109 7.284e-109
#>  [739] 4.288e-109 3.011e-109 6.696e-109 2.194e-109 2.124e-109 1.962e-109
#>  [745] 1.018e-108 1.528e-108 2.686e-109 5.497e-109 1.230e-108 2.444e-109
#>  [751] 4.584e-109 6.778e-109 2.226e-109 2.237e-109 2.045e-109 5.162e-109
#>  [757] 1.641e-108 5.358e-108 1.965e-109 1.289e-108 2.006e-109 2.801e-109
#>  [763] 2.397e-108 3.737e-109 1.116e-108 4.926e-108 3.159e-109 6.097e-109
#>  [769] 3.009e-109 2.193e-109 9.226e-109 4.565e-109 3.264e-109 1.241e-107
#>  [775] 2.043e-109 1.961e-109 3.198e-109 4.235e-109 4.529e-109 3.835e-109
#>  [781] 2.331e-109 2.813e-109 4.394e-109 3.215e-109 1.319e-108 2.070e-109
#>  [787] 3.907e-109 2.208e-109 9.445e-109 1.974e-109 4.068e-109 5.166e-109
#>  [793] 6.287e-108 1.968e-109 8.995e-109 2.139e-109 7.480e-109 1.631e-108
#>  [799] 2.670e-108 5.547e-109 2.196e-109 2.787e-108 2.818e-109 3.704e-109
#>  [805] 2.079e-109 2.526e-109 2.431e-109 2.134e-109 8.841e-109 1.844e-108
#>  [811] 2.745e-109 9.488e-109 6.535e-109 3.211e-109 4.589e-109 2.029e-109
#>  [817] 3.713e-109 4.340e-108 2.132e-109 1.657e-108 7.634e-109 8.757e-109
#>  [823] 2.087e-109 5.303e-109 5.054e-109 4.866e-109 1.991e-109 2.190e-109
#>  [829] 1.976e-109 2.227e-109 6.120e-109 6.828e-109 1.543e-108 3.338e-109
#>  [835] 1.988e-109 2.087e-109 2.570e-109 2.658e-109 6.385e-109 3.259e-109
#>  [841] 4.054e-109 3.936e-109 3.829e-109 3.200e-109 6.442e-109 3.499e-109
#>  [847] 2.026e-109 3.932e-109 3.951e-109 1.983e-109 2.112e-109 1.663e-108
#>  [853] 1.977e-109 3.934e-109 2.186e-109 1.411e-108 1.992e-109 2.301e-109
#>  [859] 3.948e-109 7.931e-109 4.661e-109 2.008e-109 1.636e-108 2.010e-109
#>  [865] 3.684e-109 5.849e-109 9.798e-109 2.201e-109 3.402e-109 6.143e-109
#>  [871] 2.073e-109 4.252e-109 9.039e-109 3.827e-109 3.017e-109 2.426e-109
#>  [877] 7.781e-109 2.922e-108 2.047e-109 1.991e-109 5.894e-109 3.363e-109
#>  [883] 9.397e-109 2.180e-109 5.269e-109 1.983e-109 4.062e-109 1.985e-109
#>  [889] 2.118e-109 2.453e-109 1.194e-108 2.397e-109 3.129e-109 2.007e-109
#>  [895] 1.974e-109 3.582e-109 2.347e-108 2.362e-108 5.089e-109 2.466e-109
#>  [901] 3.193e-108 2.943e-109 9.861e-109 5.941e-109 1.011e-108 2.603e-108
#>  [907] 4.695e-109 6.488e-109 2.758e-109 3.025e-109 2.156e-109 3.698e-109
#>  [913] 3.315e-109 8.642e-109 3.198e-107 1.961e-109 2.131e-109 1.972e-109
#>  [919] 2.631e-109 3.566e-109 3.741e-109 3.142e-109 7.676e-109 2.066e-109
#>  [925] 2.088e-109 1.989e-109 1.971e-109 1.990e-109 2.128e-109 1.663e-108
#>  [931] 3.926e-109 5.358e-109 6.626e-109 1.856e-107 3.217e-109 1.972e-109
#>  [937] 3.549e-109 3.381e-109 4.244e-109 6.189e-109 2.083e-109 6.329e-109
#>  [943] 2.005e-109 2.990e-109 5.120e-108 5.276e-109 1.814e-108 2.382e-109
#>  [949] 1.962e-109 4.167e-109 2.192e-109 5.536e-109 2.464e-109 2.857e-109
#>  [955] 2.667e-109 2.022e-109 2.520e-109 1.961e-109 1.796e-108 5.855e-109
#>  [961] 1.143e-108 3.815e-109 3.832e-109 6.422e-109 1.988e-108 1.980e-109
#>  [967] 2.996e-107 1.986e-109 2.059e-109 2.104e-109 2.185e-109 2.127e-109
#>  [973] 2.015e-109 1.841e-108 7.499e-109 3.385e-109 1.005e-107 1.972e-109
#>  [979] 2.846e-109 1.194e-108 2.476e-109 7.681e-109 2.153e-109 1.114e-108
#>  [985] 2.355e-109 1.995e-109 2.789e-109 2.303e-109 3.550e-109 7.143e-109
#>  [991] 2.862e-109 3.452e-109 5.975e-108 1.167e-108 1.967e-109 1.040e-107
#>  [997] 2.365e-109 3.958e-109 2.903e-109 3.467e-108
susie_get_niter(s)
#> [1] 4
susie_get_pip(s)
#>    [1] 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>   [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>   [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [112] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [149] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [223] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [260] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [297] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [334] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [371] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [408] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [445] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [482] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [519] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [556] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [593] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [630] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [667] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [704] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [741] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [778] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [815] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [852] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [889] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [926] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [963] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [1000] 0
susie_get_lfsr(s)
#>  [1] 0.0 0.0 0.0 0.0 0.2 0.2 0.2 0.2 0.2 0.2