Inferences From Fitted SuSiE Model

These functions access basic properties or draw inferences from a fitted susie model.

susie_get_objective(res, last_only = TRUE, warning_tol = 1e-06)

susie_get_posterior_mean(res, prior_tol = 1e-09)

susie_get_posterior_sd(res, prior_tol = 1e-09)

susie_get_niter(res)

susie_get_prior_variance(res)

susie_get_residual_variance(res)

susie_get_lfsr(res)

susie_get_posterior_samples(susie_fit, num_samples)

susie_get_cs(
  res,
  X = NULL,
  Xcorr = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  dedup = TRUE,
  squared = FALSE,
  check_symmetric = TRUE,
  n_purity = 100,
  use_rfast
)

susie_get_pip(res, prune_by_cs = FALSE, prior_tol = 1e-09)

Arguments

res

A susie fit, typically an output from susie or one of its variants. For susie_get_pip and susie_get_cs, this may instead be the posterior inclusion probability matrix, alpha.

last_only

If last_only = FALSE, return the ELBO from all iterations; otherwise return the ELBO from the last iteration only.

warning_tol

Warn if ELBO is decreasing by this tolerance level.

prior_tol

Filter out effects having estimated prior variance smaller than this threshold.

susie_fit

A susie fit, an output from susie.

num_samples

The number of draws from the posterior distribution.

X

n by p matrix of values of the p variables (covariates) in n samples. When provided, correlation between variables will be computed and used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.

Xcorr

p by p matrix of correlations between variables (covariates). When provided, it will be used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.

coverage

A number between 0 and 1 specifying desired coverage of each CS.

min_abs_corr

A "purity" threshold for the CS. Any CS that contains a pair of variables with correlation less than this threshold will be filtered out and not reported.

dedup

If dedup = TRUE, remove duplicate CSs.

squared

If squared = TRUE, report min, mean and median of squared correlation instead of the absolute correlation.

check_symmetric

If check_symmetric = TRUE, perform a check for symmetry of matrix Xcorr when Xcorr is provided (not NULL).

n_purity

The maximum number of credible set (CS) variables used in calculating the correlation (“purity”) statistics. When the number of variables included in the CS is greater than this number, the CS variables are randomly subsampled.

use_rfast

Use the Rfast package for the purity calculations. By default use_rfast = TRUE if the Rfast package is installed.

prune_by_cs

Whether or not to ignore single effects not in a reported CS when calculating PIP.

Value

susie_get_objective returns the evidence lower bound (ELBO) achieved by the fitted susie model and, optionally, at each iteration of the IBSS fitting procedure. susie_get_residual_variance returns the (estimated or fixed) residual variance parameter. susie_get_prior_variance returns the (estimated or fixed) prior variance parameters. susie_get_posterior_mean returns the posterior mean for the regression coefficients of the fitted susie model. susie_get_posterior_sd returns the posterior standard deviation for coefficients of the fitted susie model. susie_get_niter returns the number of model fitting iterations performed. susie_get_pip returns a vector containing the posterior inclusion probabilities (PIPs) for all variables. susie_get_lfsr returns a vector containing the average lfsr across variables for each single-effect, weighted by the posterior inclusion probability (alpha). susie_get_posterior_samples returns a list containing the effect sizes samples and causal status with two components: b, an num_variables x num_samples matrix of effect sizes; gamma, an num_variables x num_samplesmatrix of causal status random draws. susie_get_cs returns credible sets (CSs) from a susie fit, as well as summaries of correlation among the variables included in each CS. If desired, one can filter out CSs that do not meet a specified “purity” threshold; to do this, either X or Xcorr must be supplied. It returns a list with the following elements:

cs

A list in which each list element is a vector containing the indices of the variables in the CS.

coverage

The nominal coverage specified for each CS.

purity

If X or Xcorr iis provided), the purity of each CS.

cs_index

If X or Xcorr is provided) the index (number between 1 and L) of each reported CS in the supplied susie fit.

Examples

set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1:4] = 1
X = matrix(rnorm(n*p),nrow = n,ncol = p)
X = scale(X,center = TRUE,scale = TRUE)
y = drop(X %*% beta + rnorm(n))
s = susie(X,y,L = 10)
susie_get_objective(s)
#> [1] -1455.416
susie_get_objective(s, last_only=FALSE)
#> [1] -1927.147 -1477.646 -1455.416 -1455.416
susie_get_residual_variance(s)
#> [1] 0.9899176
susie_get_prior_variance(s)
#>  [1] 1.132714 1.049537 1.009391 0.978421 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000
susie_get_posterior_mean(s)
#>    [1]   1.063825e+00   1.023986e+00   1.004192e+00   9.886512e-01
#>    [5] -8.453415e-217 -1.417151e-213  1.201583e-214  4.561723e-216
#>    [9]  1.734347e-216 -4.657945e-217  2.154023e-216  5.614509e-216
#>   [13]  1.119569e-216  2.862909e-216 -1.153959e-216  1.523149e-218
#>   [17]  1.217335e-215  2.957793e-215  2.873810e-216 -3.406048e-216
#>   [21] -4.755227e-215 -1.673622e-215  2.574856e-215 -4.179136e-216
#>   [25]  5.353780e-216 -1.397719e-216  9.686463e-217  1.577200e-217
#>   [29]  5.922135e-217 -1.357500e-216 -2.468809e-217  8.746227e-214
#>   [33]  8.298402e-217 -6.323162e-217 -4.608892e-217  1.160039e-212
#>   [37]  3.594643e-217 -3.884377e-216 -1.543913e-216  4.805638e-216
#>   [41]  5.560123e-217  9.512582e-218 -9.188203e-217 -9.134336e-216
#>   [45] -5.319250e-216  1.607871e-213 -6.428401e-215  2.395144e-216
#>   [49] -2.447377e-217  3.164270e-215 -1.420230e-215 -2.330788e-217
#>   [53]  1.873140e-217 -1.018681e-216  1.511235e-216  4.938063e-216
#>   [57]  7.497066e-217 -5.945768e-217 -6.349095e-216  1.211021e-215
#>   [61]  7.046490e-217 -2.809440e-215  3.566071e-217 -6.749748e-217
#>   [65] -2.882037e-217  5.030366e-217  1.203192e-215  1.894047e-216
#>   [69] -1.304107e-212  5.338713e-217  8.686288e-216  1.800492e-214
#>   [73]  2.853655e-217 -5.237734e-215 -3.620615e-216 -4.332929e-217
#>   [77]  1.144477e-216  1.609066e-215  4.164065e-214 -2.362128e-215
#>   [81] -7.197737e-217  6.354672e-217  4.804369e-216 -1.580831e-217
#>   [85]  3.802540e-216 -3.296370e-216 -3.714979e-218  6.676895e-217
#>   [89] -3.205994e-216  2.746267e-216 -4.360117e-216 -5.908594e-215
#>   [93]  6.768710e-217  1.817145e-216 -4.795225e-217  8.461871e-217
#>   [97] -7.122763e-217  1.489105e-215 -8.621147e-216 -1.363945e-216
#>  [101]  1.711542e-217  8.756724e-216 -1.906718e-214 -1.911585e-216
#>  [105]  3.787328e-216  3.310293e-214 -1.766389e-216 -5.336149e-216
#>  [109]  6.925433e-217  2.332834e-216  4.733749e-217  2.230102e-215
#>  [113] -6.674221e-216  3.677889e-218  5.456026e-216  1.432486e-216
#>  [117] -2.773765e-217  1.210457e-216 -9.883160e-217  7.835604e-216
#>  [121]  4.984242e-217 -1.352101e-216  8.744269e-217 -1.518912e-215
#>  [125]  9.001101e-217 -7.813814e-217 -1.198923e-215  4.466498e-216
#>  [129]  6.151252e-217  3.081636e-215  2.764383e-215  5.335474e-217
#>  [133] -4.453473e-217 -3.049091e-214  3.333238e-217  4.059981e-217
#>  [137]  6.812472e-218 -3.098747e-216  5.098449e-216  5.813210e-217
#>  [141] -5.065428e-217  1.005287e-216  3.097971e-217 -5.377436e-217
#>  [145] -1.027776e-216 -1.438676e-215  1.518813e-214  1.169116e-216
#>  [149]  4.908786e-216 -5.611277e-216  3.415806e-217  1.954231e-216
#>  [153]  2.989363e-216  5.507984e-217  2.090598e-217 -2.407923e-215
#>  [157] -1.364342e-216 -2.111521e-216 -1.135603e-215 -1.625525e-216
#>  [161] -9.847530e-217 -9.760085e-215 -1.406766e-214  4.288097e-217
#>  [165] -3.570655e-216  7.880394e-215 -6.005368e-217  1.886486e-217
#>  [169]  4.400035e-216  1.916333e-214  2.074352e-216  7.297415e-217
#>  [173] -3.007515e-216 -2.471356e-216 -5.288808e-216 -7.911513e-217
#>  [177] -2.158625e-216  2.344281e-215 -3.252106e-216  7.732715e-215
#>  [181]  1.056887e-217 -8.223976e-216 -1.835361e-216  2.387140e-216
#>  [185]  1.495170e-215  2.547701e-217 -3.763749e-215 -1.203568e-217
#>  [189] -1.795721e-216 -2.285112e-215 -4.364788e-217 -1.153756e-216
#>  [193]  7.928309e-216  2.672654e-217  3.477151e-217 -1.766856e-216
#>  [197] -1.096530e-216 -2.950116e-216 -1.045740e-216  1.260164e-216
#>  [201]  5.503766e-216 -2.288041e-214  1.017713e-216 -4.558442e-215
#>  [205]  3.710221e-215 -7.383942e-217  3.731530e-216  1.204640e-215
#>  [209] -1.100836e-217 -2.296265e-215 -8.247191e-217  3.225623e-218
#>  [213] -1.082281e-216 -1.358979e-213 -4.782181e-216 -1.531988e-217
#>  [217] -1.732409e-214 -5.208158e-216  1.604457e-215 -1.356741e-217
#>  [221] -8.577181e-217 -8.091139e-217 -8.123132e-217  7.593253e-217
#>  [225] -6.029480e-216  1.186547e-216  2.929143e-216 -1.593338e-215
#>  [229] -6.386229e-216 -4.185052e-213  2.568616e-216 -1.211651e-216
#>  [233]  2.457526e-216  8.303359e-216  1.862521e-215 -3.387281e-216
#>  [237]  1.660586e-217 -9.643088e-216  5.009111e-216 -1.762933e-215
#>  [241]  6.913823e-217 -9.056835e-218  8.903106e-216  4.776681e-216
#>  [245]  7.994141e-217  3.788216e-215  4.257166e-217 -1.321800e-216
#>  [249]  5.918242e-218  1.740664e-216 -5.330869e-217  1.515586e-217
#>  [253] -1.208393e-216 -1.322063e-216 -4.289292e-217  2.334949e-216
#>  [257]  4.526367e-217 -7.226022e-217  4.873802e-217  4.105103e-217
#>  [261] -1.126484e-216 -4.643134e-217 -5.811264e-217 -6.008250e-216
#>  [265]  2.122859e-215  2.595463e-217 -5.697126e-216 -1.808049e-216
#>  [269]  3.315668e-216  2.289490e-216 -1.476663e-216 -1.635013e-217
#>  [273] -1.439169e-212 -2.937969e-217  2.574745e-213 -1.464897e-216
#>  [277]  1.349383e-216  6.644532e-219 -3.520317e-216  2.706193e-216
#>  [281] -1.906314e-215 -8.881822e-216  8.199132e-217 -3.244941e-216
#>  [285]  1.364082e-218 -2.395516e-218 -1.865011e-216  2.516864e-217
#>  [289] -8.450955e-218  8.602545e-217 -1.163393e-217  1.603950e-214
#>  [293] -4.195476e-217  2.919372e-217 -2.790204e-216  4.070217e-216
#>  [297] -4.880283e-215 -8.955565e-216  5.084957e-217  3.924742e-216
#>  [301] -3.890152e-218 -9.458991e-217  9.715167e-213 -1.212440e-216
#>  [305]  7.153850e-217 -3.993466e-216  7.714205e-216 -2.997400e-216
#>  [309] -2.600955e-215  2.138881e-216  7.738079e-217 -6.063984e-219
#>  [313]  1.043387e-216  1.165383e-216  4.819641e-216  1.804675e-215
#>  [317] -4.074579e-216  3.446046e-216  3.054389e-212 -1.346183e-216
#>  [321]  4.548667e-217 -1.006963e-215  9.171383e-218  6.648108e-216
#>  [325] -6.458434e-217  3.211285e-215 -2.142650e-216  6.247989e-217
#>  [329] -1.096480e-215 -1.588108e-216  2.121316e-217  1.633976e-215
#>  [333]  6.041246e-216  6.387666e-218  1.307924e-217  6.461389e-215
#>  [337] -8.266950e-216  4.563527e-216  1.617164e-217 -1.132624e-216
#>  [341]  5.228956e-216 -2.475274e-217  3.229269e-216 -7.204623e-216
#>  [345]  8.456075e-217  1.557069e-214  4.271069e-216 -1.606357e-216
#>  [349]  1.617610e-217  8.311646e-216  3.267844e-216 -3.657800e-216
#>  [353]  6.368637e-216 -4.042059e-215  1.207844e-215 -6.174982e-217
#>  [357] -9.995589e-215 -1.110565e-216 -7.324852e-214 -1.242884e-217
#>  [361] -3.295606e-215  1.705821e-215  2.573258e-218 -2.298261e-214
#>  [365] -9.716227e-215 -2.964152e-216 -1.208990e-216  1.434601e-216
#>  [369]  3.266586e-215  5.341370e-217 -7.211391e-216  1.982229e-217
#>  [373]  8.752747e-217 -4.584595e-215  1.291725e-216  2.719277e-218
#>  [377] -5.300150e-216  5.186464e-216 -2.882413e-216  8.720335e-216
#>  [381] -5.496038e-217 -1.586379e-216  8.672701e-214  3.471245e-217
#>  [385] -1.459058e-214  2.768301e-217 -1.154042e-216 -2.106037e-216
#>  [389] -7.965320e-218  4.590662e-217  1.158414e-214  1.433776e-216
#>  [393] -7.785828e-217 -6.112129e-217  7.356650e-217 -5.242807e-216
#>  [397] -1.568227e-213  5.712025e-217  7.236770e-218  1.538676e-215
#>  [401]  1.577647e-216 -2.140943e-216  2.737315e-216 -1.428319e-215
#>  [405]  3.914283e-217 -1.835093e-216 -3.290047e-216 -5.780672e-217
#>  [409] -6.155320e-215 -3.402409e-216 -3.677303e-216  1.315248e-216
#>  [413]  4.344609e-216  4.521614e-217 -8.248374e-215 -1.721434e-216
#>  [417] -3.842478e-216  1.385441e-215 -1.799752e-216 -8.137429e-214
#>  [421]  7.291286e-216  2.836489e-216  2.883507e-216 -4.488321e-217
#>  [425] -3.740672e-216 -6.679845e-216  5.023562e-217 -1.755977e-215
#>  [429]  4.303305e-216 -1.937935e-216 -3.806384e-217  3.832329e-217
#>  [433]  1.657362e-214  1.955964e-216  2.881432e-216 -1.846103e-216
#>  [437]  2.414691e-216  1.003295e-216  5.749078e-217 -1.897208e-216
#>  [441]  2.137203e-216  4.608212e-217 -1.843146e-216 -3.532281e-216
#>  [445] -1.095628e-215 -4.184547e-217 -2.491527e-217 -9.836118e-216
#>  [449]  2.717185e-217  1.109231e-216 -4.323917e-214 -2.981151e-216
#>  [453]  4.333319e-216  1.053529e-216  9.364600e-216  1.832127e-216
#>  [457]  4.113126e-215 -3.087182e-216  3.701262e-215 -3.374066e-216
#>  [461]  1.059537e-215  7.329713e-216 -2.977083e-216  1.622697e-216
#>  [465] -2.115836e-217  7.261185e-217 -4.885790e-217  5.658757e-216
#>  [469]  1.701756e-215  2.197174e-215 -7.433406e-216  1.298488e-216
#>  [473] -2.672121e-216  3.721317e-218  3.769903e-216 -1.537263e-216
#>  [477] -3.986902e-216  1.566943e-215 -9.148298e-217  1.835966e-216
#>  [481] -1.165385e-215 -4.961478e-215  1.089097e-215 -6.668232e-217
#>  [485]  2.001819e-217 -1.649975e-213 -1.703184e-216 -8.430521e-217
#>  [489] -6.197190e-216 -3.554246e-215 -8.757716e-217 -6.050441e-217
#>  [493]  1.406671e-217 -5.778929e-216  5.221640e-218 -1.207028e-216
#>  [497]  1.236405e-216  9.301018e-217 -7.571449e-217  1.005966e-215
#>  [501]  2.852615e-217 -4.659666e-215  2.067395e-215  3.179569e-217
#>  [505]  4.636637e-217 -8.934884e-217 -1.908562e-217  2.096515e-216
#>  [509] -4.628876e-216  3.938280e-216 -1.674374e-215 -5.240854e-215
#>  [513]  2.083724e-216  4.788554e-218 -1.149079e-214 -1.859830e-215
#>  [517]  3.040875e-216  1.632337e-216 -1.175233e-214 -7.497011e-217
#>  [521]  5.920520e-218 -4.512253e-216 -2.940845e-218  7.981814e-218
#>  [525]  9.944307e-217  1.933512e-214  4.159091e-217 -6.898471e-217
#>  [529]  3.398976e-215 -2.642470e-216  7.784348e-216  4.274897e-216
#>  [533] -4.109507e-214 -1.987659e-216  5.944978e-217  1.339260e-216
#>  [537] -4.732177e-216  1.076050e-216 -4.003456e-216 -5.017297e-217
#>  [541]  1.605572e-218  2.404383e-216 -6.844595e-219 -5.627016e-216
#>  [545]  2.535615e-216 -1.176805e-216 -1.030738e-216  5.228333e-216
#>  [549]  1.363719e-213  1.781139e-216  1.264844e-216 -5.776243e-217
#>  [553] -9.490408e-217 -1.345139e-214 -2.059103e-211  8.441850e-217
#>  [557]  1.580856e-217  3.422738e-216  1.414511e-216  4.247428e-216
#>  [561]  8.564411e-217 -4.448577e-216  4.525635e-217  8.186769e-217
#>  [565]  3.359225e-214 -4.654940e-217 -4.593802e-216 -4.402906e-215
#>  [569]  6.124091e-216  3.391552e-217 -2.741471e-215 -6.880474e-216
#>  [573] -6.872535e-217 -7.491577e-217  3.127045e-216  3.579982e-217
#>  [577]  4.039908e-216  2.167478e-216  2.650572e-216 -9.423535e-218
#>  [581]  2.644900e-216  4.085465e-216 -2.096888e-216  4.238986e-216
#>  [585] -2.250166e-216 -1.455199e-215 -2.185123e-217  1.106997e-216
#>  [589]  2.801556e-217  6.746333e-216  6.531352e-216 -4.162829e-216
#>  [593]  4.141767e-216  4.829666e-214 -1.513988e-216 -3.472103e-216
#>  [597]  2.116471e-216  1.161767e-217  1.416050e-216  2.052074e-215
#>  [601] -6.662535e-218  1.022159e-216  2.033348e-215  4.219852e-217
#>  [605]  7.623675e-216  1.115567e-216  1.296443e-217 -1.332073e-215
#>  [609]  3.169995e-217  1.670617e-215  3.153569e-217  3.557483e-211
#>  [613] -5.449886e-217 -6.054374e-216 -6.966711e-218 -1.567848e-216
#>  [617] -6.441374e-217  2.593736e-215 -1.153667e-215  4.526473e-217
#>  [621]  9.337272e-216  9.792551e-217 -4.076786e-217 -1.222130e-214
#>  [625] -2.757835e-215 -1.959382e-215 -1.570916e-216  6.619632e-216
#>  [629] -1.548267e-216 -1.437344e-217 -1.592390e-215  4.920118e-217
#>  [633] -4.120101e-216 -6.558110e-217  1.317222e-216 -8.103479e-215
#>  [637]  4.865119e-216 -1.229402e-216  3.073763e-217 -9.862812e-216
#>  [641]  5.901298e-215  7.798518e-218 -4.146195e-213 -6.453182e-216
#>  [645]  2.409365e-217 -5.737021e-217 -2.461369e-217 -1.445973e-216
#>  [649]  2.617054e-216 -1.818298e-216  5.528300e-217  4.256421e-217
#>  [653]  2.483329e-216 -1.257865e-217  6.052209e-216 -9.544615e-217
#>  [657] -9.766109e-217  6.115016e-216 -2.390106e-217  6.727624e-217
#>  [661]  9.469454e-216  2.349519e-217 -7.239949e-215  1.219901e-216
#>  [665]  1.427140e-216  6.119765e-217 -3.008555e-216 -6.908254e-217
#>  [669] -9.117257e-217  1.275996e-216  1.378863e-215 -5.334567e-218
#>  [673]  1.853146e-216 -3.208341e-217  4.037935e-217  6.813842e-217
#>  [677]  1.503606e-217 -1.191157e-216 -1.132159e-217  7.253271e-216
#>  [681]  6.948190e-217  7.273466e-216 -3.307361e-216 -8.054776e-217
#>  [685] -2.168044e-216 -1.693032e-213 -1.422252e-216 -2.857733e-218
#>  [689]  1.530881e-213  1.589734e-216  3.052321e-215  3.558237e-216
#>  [693]  2.195097e-217  1.192190e-216 -4.439871e-217 -8.613446e-217
#>  [697] -4.416606e-219  1.758333e-217  2.841791e-216 -9.494769e-217
#>  [701] -1.482531e-215 -1.283965e-215 -1.953802e-216  1.048496e-216
#>  [705]  1.399608e-216  1.026791e-215  9.450733e-216  8.969974e-217
#>  [709] -3.023248e-216 -3.223879e-215  6.141157e-215  1.178430e-214
#>  [713]  8.438564e-217 -3.838199e-216  1.761639e-216 -4.924419e-215
#>  [717]  4.114684e-216 -1.508014e-216 -8.554463e-217 -7.505987e-216
#>  [721]  5.486081e-217  1.231488e-216 -6.691354e-217  5.906285e-215
#>  [725] -2.106497e-216  2.606913e-217 -8.026654e-215 -9.365267e-214
#>  [729]  7.215045e-219  1.640751e-216 -1.805976e-216  2.752812e-213
#>  [733] -1.465475e-216 -6.659034e-216 -7.030432e-216 -7.845623e-217
#>  [737] -3.814659e-216  7.516719e-216 -2.882546e-216  1.414592e-216
#>  [741] -6.470588e-216 -5.246480e-217  4.287105e-217  3.422964e-218
#>  [745]  1.364086e-215 -2.817717e-215 -1.072754e-216 -4.543846e-216
#>  [749] -1.911205e-215  8.147229e-217  3.263967e-216 -6.613132e-216
#>  [753]  5.651580e-217 -5.788865e-217 -2.997883e-217 -4.056275e-216
#>  [757] -3.204553e-215  2.776173e-214 -6.440735e-218  2.078325e-215
#>  [761]  2.183368e-217  1.193334e-216 -6.353777e-215  2.216590e-216
#>  [765] -1.604963e-215  2.376875e-214  1.572423e-216  5.473849e-216
#>  [769] -1.411755e-216  5.223751e-217  1.144416e-215  3.239200e-216
#>  [773] -1.685752e-216  1.325392e-213 -2.970639e-217  1.138729e-218
#>  [777]  1.614578e-216 -2.817049e-216 -3.192132e-216  2.331554e-216
#>  [781]  6.888753e-217  1.205403e-216  3.018058e-216  1.632304e-216
#>  [785] -2.164260e-215 -3.443764e-217  2.416939e-216  5.417462e-217
#>  [789] -1.193062e-215  1.159254e-217 -2.610390e-216  4.060786e-216
#>  [793] -3.732159e-214  8.425381e-218  1.093886e-215 -4.506839e-217
#>  [797] -7.879620e-216 -3.166866e-215 -7.731182e-215 -4.618741e-216
#>  [801]  5.272546e-217  8.357745e-215 -1.211012e-216  2.179007e-216
#>  [805] -3.601223e-217 -9.032681e-217  8.011898e-217  4.437618e-217
#>  [809] -1.060856e-215  3.954312e-215  1.134040e-216  1.202711e-215
#>  [813]  6.195694e-216  1.627931e-216 -3.270155e-216 -2.701018e-217
#>  [817] -2.189186e-216  1.881536e-214  4.401764e-217 -3.260919e-215
#>  [821] -8.171759e-216 -1.043047e-215 -3.726179e-217 -4.258362e-216
#>  [825] -3.902257e-216 -3.641678e-216 -1.758922e-217 -5.188605e-217
#>  [829]  1.245528e-217  5.659281e-217  5.509489e-216 -6.699349e-216
#>  [833]  2.866548e-215  1.766867e-216  1.683123e-217  3.720970e-217
#>  [837] -9.503680e-217 -1.043348e-216  5.944579e-216  1.680247e-216
#>  [841] -2.593312e-216  2.451304e-216  2.324848e-216  1.615977e-216
#>  [845]  6.039326e-216  1.945627e-216 -2.620956e-217 -2.445990e-216
#>  [849]  2.468816e-216  1.497642e-217  4.117312e-217  3.279991e-215
#>  [853]  1.271709e-217  2.448701e-216  5.139089e-217 -2.443482e-215
#>  [857]  1.789110e-217  6.552099e-217  2.466118e-216 -8.745182e-216
#>  [861] -3.365772e-216 -2.214994e-217  3.186699e-215 -2.272083e-217
#>  [865]  2.155556e-216  5.080271e-216 -1.273605e-215  5.331571e-217
#>  [869] -1.837609e-216  5.547521e-216 -3.494234e-217 -2.837869e-216
#>  [873] -1.103409e-215 -2.322071e-216  1.421035e-216  7.949849e-217
#>  [877]  8.452882e-216  9.112561e-215  3.034474e-217 -1.781216e-217
#>  [881]  5.150989e-216 -1.793767e-216  1.182358e-215  5.064350e-217
#>  [885] -4.208904e-216 -1.511314e-217 -2.603377e-216 -1.594495e-217
#>  [889]  4.201508e-217  8.253568e-217 -1.810559e-215  7.633345e-217
#>  [893] -1.539724e-216 -2.191629e-217 -1.163739e-217  2.038968e-216
#>  [897]  6.114693e-215 -6.187612e-215 -3.951902e-216 -8.386076e-217
#>  [901]  1.071789e-214  1.342111e-216  1.288173e-215  5.224519e-216
#>  [905] -1.347736e-215 -7.381100e-215 -3.411162e-216  6.116849e-216
#>  [909] -1.147600e-216  1.428938e-216 -4.739472e-217 -2.171298e-216
#>  [913] -1.741950e-216  1.018736e-215  7.904034e-213  2.302359e-218
#>  [917] -4.384965e-217 -1.061610e-217 -1.014233e-216 -2.020908e-216
#>  [921] -2.222052e-216 -1.554296e-216  8.252240e-216  3.375281e-217
#>  [925] -3.747266e-217  1.712991e-217 -1.045983e-217 -1.747777e-217
#>  [929] -4.349827e-217 -3.280187e-215  2.438763e-216 -4.339318e-216
#>  [933]  6.350826e-216 -2.825334e-213  1.634378e-216  1.070040e-217
#>  [937] -2.001689e-216 -1.814303e-216  2.827264e-216 -5.622449e-216
#>  [941] -3.668116e-217  5.850695e-216 -2.138383e-217  1.392341e-216
#>  [945] -2.552108e-214 -4.220075e-216  3.835893e-215  7.466543e-217
#>  [949]  3.763897e-218  2.732076e-216 -5.216961e-217 -4.601935e-216
#>  [953]  8.364991e-217 -1.252098e-216  1.053079e-216 -2.537830e-217
#>  [957] -8.971744e-217  2.474750e-218 -3.767368e-215 -5.089541e-216
#>  [961]  1.676999e-215  2.307456e-216  2.327567e-216 -6.005895e-216
#>  [965] -4.528749e-215  1.421242e-217 -6.984570e-213  1.627735e-217
#>  [969] -3.267888e-217 -3.994392e-217  5.126129e-217 -4.328507e-217
#>  [973]  2.386807e-217 -3.940193e-215 -7.916159e-216 -1.818064e-216
#>  [977] -8.942952e-214  1.075354e-217  1.240179e-216  1.810859e-215
#>  [981]  8.493123e-217 -8.260577e-216  4.701119e-217 -1.600956e-215
#>  [985] -7.170516e-217  1.894373e-217 -1.180732e-216 -6.570960e-217
#>  [989] -2.003210e-216 -7.259567e-216 -1.256882e-216  1.892978e-216
#>  [993] -3.396743e-214 -1.739134e-215  7.822050e-218  9.530524e-214
#>  [997] -7.282605e-217  2.477154e-216 -1.299976e-216  1.246133e-214
susie_get_posterior_sd(s)
#>    [1]  3.146494e-02  3.146385e-02  3.146326e-02  3.146277e-02 2.471927e-109
#>    [6] 1.285636e-107 3.399189e-108 5.508877e-109 3.308551e-109 2.150000e-109
#>   [11] 3.682613e-109 6.184568e-109 2.730881e-109 4.271947e-109 2.763642e-109
#>   [16] 1.961004e-109 9.552288e-109 1.569790e-108 4.280621e-109 4.691272e-109
#>   [21] 2.042720e-108 1.142114e-108 1.453123e-108 5.248127e-109 6.022398e-109
#>   [26] 2.995420e-109 2.587348e-109 1.984895e-109 2.248339e-109 2.957375e-109
#>   [31] 2.018592e-109 9.934418e-108 2.457562e-109 2.281847e-109 2.146493e-109
#>   [36] 3.914998e-107 2.078700e-109 5.040843e-109 3.132686e-109 5.670599e-109
#>   [41] 2.218948e-109 1.969635e-109 2.540383e-109 8.127608e-109 6.000683e-109
#>   [46] 1.375137e-107 2.412336e-108 3.888876e-109 2.017628e-109 1.629827e-108
#>   [51] 1.041631e-108 2.012511e-109 1.994950e-109 2.634748e-109 3.102153e-109
#>   [56] 5.757020e-109 2.384542e-109 2.250284e-109 6.625328e-109 9.524888e-109
#>   [61] 2.344430e-109 1.525462e-108 2.076962e-109 2.318453e-109 2.038559e-109
#>   [66] 2.177647e-109 9.489752e-109 3.453151e-109 4.163831e-107 2.201424e-109
#>   [71] 7.900908e-109 4.237004e-108 2.037105e-109 2.154743e-108 4.850201e-109
#>   [76] 2.126927e-109 2.754594e-109 1.117201e-108 6.669216e-108 1.384880e-108
#>   [81] 2.357804e-109 2.284521e-109 5.669769e-109 1.985002e-109 4.982226e-109
#>   [86] 4.608657e-109 1.962133e-109 2.312138e-109 4.539774e-109 4.178438e-109
#>   [91] 5.372594e-109 2.302857e-108 2.320101e-109 3.383861e-109 2.160047e-109
#>   [96] 2.472638e-109 2.351160e-109 1.069688e-108 7.867529e-109 2.963479e-109
#>  [101] 1.989099e-109 7.936879e-109 4.371043e-108 3.468863e-109 4.971279e-109
#>  [106] 5.892852e-108 3.337782e-109 6.011317e-109 2.333783e-109 3.836167e-109
#>  [111] 2.155526e-109 1.341112e-108 6.813541e-109 1.962106e-109 6.086362e-109
#>  [116] 3.028217e-109 2.033081e-109 2.817545e-109 2.605946e-109 7.456025e-109
#>  [121] 2.174157e-109 2.952263e-109 2.498861e-109 1.081649e-108 2.522842e-109
#>  [126] 2.413171e-109 9.470813e-109 5.444809e-109 2.267367e-109 1.606021e-108
#>  [131] 1.511788e-108 2.201171e-109 2.135362e-109 5.636925e-108 2.063186e-109
#>  [136] 2.108389e-109 1.965332e-109 4.457172e-109 5.860443e-109 2.239410e-109
#>  [141] 2.180313e-109 2.622035e-109 2.050016e-109 2.204463e-109 2.643389e-109
#>  [146] 1.049204e-108 3.862446e-108 2.778106e-109 5.737995e-109 6.182549e-109
#>  [151] 2.067996e-109 3.506930e-109 4.371888e-109 2.214789e-109 2.002663e-109
#>  [156] 1.399800e-108 2.963854e-109 3.645605e-109 9.186285e-109 3.208488e-109
#>  [161] 2.602571e-109 3.033887e-108 3.704501e-108 2.123826e-109 4.813514e-109
#>  [166] 2.698126e-108 2.255213e-109 1.995053e-109 5.399772e-109 4.383017e-108
#>  [171] 3.613079e-109 2.366671e-109 4.386127e-109 3.952789e-109 6.200176e-109
#>  [176] 2.422064e-109 3.686608e-109 1.379030e-108 4.574988e-109 2.670244e-108
#>  [181] 1.971698e-109 7.661598e-109 3.400332e-109 3.882126e-109 1.072130e-108
#>  [186] 2.022203e-109 1.794651e-108 1.974911e-109 3.364448e-109 1.359488e-108
#>  [191] 2.129147e-109 2.763449e-109 7.505488e-109 2.028118e-109 2.071619e-109
#>  [196] 3.338207e-109 2.708854e-109 4.341027e-109 2.660471e-109 2.864915e-109
#>  [201] 6.116058e-109 4.825547e-108 2.633828e-109 1.995535e-108 1.780449e-108
#>  [206] 2.374398e-109 4.930982e-109 9.496173e-109 1.972618e-109 1.363189e-108
#>  [211] 2.452820e-109 1.961936e-109 2.695276e-109 1.257214e-107 5.655192e-109
#>  [216] 1.983550e-109 4.149143e-108 5.930427e-109 1.115405e-108 1.978695e-109
#>  [221] 2.483325e-109 2.438484e-109 2.441428e-109 2.393201e-109 6.436311e-109
#>  [226] 2.794738e-109 4.324483e-109 1.111064e-108 6.647027e-109 2.285459e-107
#>  [231] 4.034097e-109 2.818684e-109 3.941235e-109 7.703096e-109 1.212616e-108
#>  [236] 4.677200e-109 1.987466e-109 8.379217e-109 5.803000e-109 1.175870e-108
#>  [241] 2.332765e-109 1.968816e-109 8.011232e-109 5.651574e-109 2.429619e-109
#>  [246] 1.801127e-108 2.121700e-109 2.923521e-109 1.964217e-109 3.314322e-109
#>  [251] 2.200811e-109 1.983072e-109 2.815587e-109 2.923770e-109 2.124115e-109
#>  [256] 3.837963e-109 2.140535e-109 2.360317e-109 2.165875e-109 2.111393e-109
#>  [261] 2.737426e-109 2.149134e-109 2.239244e-109 6.423610e-109 1.304686e-108
#>  [266] 2.024436e-109 6.235265e-109 3.375623e-109 4.623237e-109 3.799261e-109
#>  [271] 3.069757e-109 1.986663e-109 4.385404e-107 2.041456e-109 1.766488e-107
#>  [276] 3.058708e-109 2.949684e-109 1.960820e-109 4.776357e-109 4.146011e-109
#>  [281] 1.228493e-108 8.000455e-109 2.448400e-109 4.569529e-109 1.960958e-109
#>  [286] 1.961340e-109 3.427065e-109 2.020792e-109 1.967778e-109 2.485680e-109
#>  [291] 1.973991e-109 3.978845e-108 2.117489e-109 2.040488e-109 4.213814e-109
#>  [296] 5.172211e-109 2.072239e-108 8.037749e-109 2.181801e-109 5.069571e-109
#>  [301] 1.962264e-109 2.565859e-109 3.565816e-107 2.819448e-109 2.353912e-109
#>  [306] 5.118236e-109 7.390870e-109 4.378191e-109 1.461316e-108 3.669451e-109
#>  [311] 2.406299e-109 1.960811e-109 2.658233e-109 2.774544e-109 5.679782e-109
#>  [316] 1.191383e-108 5.175263e-109 4.721172e-109 6.510489e-107 2.946652e-109
#>  [321] 2.142126e-109 8.585721e-109 1.969106e-109 6.798568e-109 2.293355e-109
#>  [326] 1.643245e-108 3.672729e-109 2.275495e-109 9.006967e-109 3.173811e-109
#>  [331] 2.003874e-109 1.126865e-108 6.443343e-109 1.964783e-109 1.977441e-109
#>  [336] 2.419143e-108 7.684087e-109 5.510086e-109 1.986112e-109 2.743284e-109
#>  [341] 5.943627e-109 2.018884e-109 4.557571e-109 7.112437e-109 2.472101e-109
#>  [346] 3.915116e-108 5.311611e-109 3.190740e-109 1.986124e-109 7.707419e-109
#>  [351] 4.586966e-109 4.877385e-109 6.636754e-109 1.867043e-108 9.510358e-109
#>  [356] 2.269356e-109 3.073745e-108 2.722239e-109 9.034109e-108 1.975840e-109
#>  [361] 1.667085e-108 1.154373e-108 1.961427e-109 4.837218e-108 3.026415e-108
#>  [366] 4.352080e-109 2.816148e-109 3.030210e-109 1.658912e-108 2.201633e-109
#>  [371] 7.116189e-109 1.999146e-109 2.499652e-109 2.001859e-108 2.894946e-109
#>  [376] 1.961509e-109 5.988647e-109 5.916636e-109 4.287458e-109 7.918311e-109
#>  [381] 2.213839e-109 3.172205e-109 9.889626e-108 2.071268e-109 3.778918e-108
#>  [386] 2.032807e-109 2.763722e-109 3.640846e-109 1.966999e-109 2.145138e-109
#>  [391] 3.331931e-108 3.029432e-109 2.410628e-109 2.264149e-109 2.371958e-109
#>  [396] 5.952405e-109 1.356966e-107 2.231183e-109 1.965915e-109 1.089522e-108
#>  [401] 3.164091e-109 3.671244e-109 4.171207e-109 1.044957e-108 2.098818e-109
#>  [406] 3.400090e-109 4.603830e-109 2.236753e-109 2.355354e-108 4.688545e-109
#>  [411] 4.891602e-109 2.917299e-109 5.362009e-109 2.140195e-109 2.766566e-108
#>  [416] 3.296740e-109 5.010888e-109 1.027228e-108 3.368102e-109 9.557973e-108
#>  [421] 7.160371e-109 4.250883e-109 4.288327e-109 2.137828e-109 4.937600e-109
#>  [426] 6.816760e-109 2.177132e-109 1.173268e-108 5.333744e-109 3.492407e-109
#>  [431] 2.091891e-109 2.093545e-109 4.050401e-108 3.508472e-109 4.286679e-109
#>  [436] 3.410035e-109 3.905325e-109 2.620145e-109 2.234178e-109 3.455986e-109
#>  [441] 3.667991e-109 2.146403e-109 3.407361e-109 4.785207e-109 9.003032e-109
#>  [446] 2.116747e-109 2.019622e-109 8.473155e-109 2.030284e-109 2.720966e-109
#>  [451] 6.805930e-108 4.365443e-109 5.354294e-109 2.667884e-109 8.242227e-109
#>  [456] 3.397412e-109 1.885157e-108 4.448206e-109 1.778064e-108 4.667273e-109
#>  [461] 8.835032e-109 7.181547e-109 4.362247e-109 3.205872e-109 2.003658e-109
#>  [466] 2.363448e-109 2.166769e-109 6.211746e-109 1.152841e-108 1.330023e-108
#>  [471] 7.238453e-109 2.901375e-109 4.118317e-109 1.962138e-109 4.958719e-109
#>  [476] 3.126476e-109 5.113602e-109 1.100704e-108 2.536637e-109 3.400879e-109
#>  [481] 9.320962e-109 2.091218e-108 8.972815e-109 2.311389e-109 1.999266e-109
#>  [486] 1.394204e-107 3.280019e-109 2.469737e-109 6.536010e-109 1.738520e-108
#>  [491] 2.500118e-109 2.258955e-109 1.980020e-109 6.285192e-109 1.963456e-109
#>  [496] 2.814276e-109 2.842283e-109 2.550984e-109 2.391235e-109 8.580938e-109
#>  [501] 2.037052e-109 2.019921e-108 1.285520e-108 2.054484e-109 2.148462e-109
#>  [506] 2.516648e-109 1.995842e-109 3.632533e-109 5.553738e-109 5.079183e-109
#>  [511] 1.142398e-108 2.155452e-108 3.621298e-109 1.963051e-109 3.317234e-108
#>  [516] 1.211634e-108 4.412183e-109 3.214787e-109 3.358272e-108 2.384540e-109
#>  [521] 1.964221e-109 5.475660e-109 1.961628e-109 1.967024e-109 2.611740e-109
#>  [526] 4.404296e-108 2.115025e-109 2.331421e-109 1.695933e-108 4.094125e-109
#>  [531] 7.428569e-109 5.314242e-109 6.621991e-108 3.536629e-109 2.250302e-109
#>  [536] 2.940088e-109 5.622246e-109 2.689332e-109 5.125314e-109 2.176657e-109
#>  [541] 1.961029e-109 3.896655e-109 1.960822e-109 6.192239e-109 4.006212e-109
#>  [546] 2.785443e-109 2.646204e-109 5.943231e-109 1.259551e-107 3.351201e-109
#>  [551] 2.869370e-109 2.236388e-109 2.568822e-109 3.615122e-108 1.764030e-106
#>  [556] 2.470793e-109 1.985003e-109 4.703764e-109 3.011271e-109 5.295337e-109
#>  [561] 2.482142e-109 5.432692e-109 2.140484e-109 2.447262e-109 5.939726e-108
#>  [566] 2.149781e-109 5.530340e-109 1.957572e-108 6.492692e-109 2.066573e-109
#>  [571] 1.504796e-108 6.930928e-109 2.329456e-109 2.384048e-109 4.479072e-109
#>  [576] 2.077816e-109 5.150941e-109 3.694288e-109 4.100744e-109 1.969471e-109
#>  [581] 4.096118e-109 5.182882e-109 3.632821e-109 5.289517e-109 3.765604e-109
#>  [586] 1.055951e-108 2.006426e-109 2.718836e-109 2.034469e-109 6.854754e-109
#>  [591] 6.731322e-109 5.236810e-109 5.222168e-109 7.223496e-108 3.104729e-109
#>  [596] 4.740581e-109 3.649926e-109 1.973963e-109 3.012732e-109 1.280185e-108
#>  [601] 1.965136e-109 2.638055e-109 1.273639e-108 2.119148e-109 7.341995e-109
#>  [606] 2.727012e-109 1.977153e-109 1.004815e-108 2.053951e-109 1.140961e-108
#>  [611] 2.053042e-109 2.344434e-106 2.210178e-109 6.451182e-109 1.965540e-109
#>  [616] 3.154978e-109 2.291899e-109 1.459054e-108 9.268156e-109 2.140541e-109
#>  [621] 8.228925e-109 2.597366e-109 2.109502e-109 3.430807e-108 1.509792e-108
#>  [626] 1.247513e-108 3.157832e-109 6.782216e-109 3.136744e-109 1.980872e-109
#>  [631] 1.110694e-108 2.169335e-109 5.207075e-109 2.301895e-109 2.919175e-109
#>  [636] 2.739790e-108 5.709534e-109 2.835608e-109 2.048677e-109 8.486099e-109
#>  [641] 2.301289e-108 1.966741e-109 2.274199e-107 6.686016e-109 2.015936e-109
#>  [646] 2.233201e-109 2.018258e-109 3.040916e-109 4.073318e-109 3.384905e-109
#>  [651] 2.216407e-109 2.121812e-109 3.962775e-109 1.976202e-109 6.449889e-109
#>  [656] 2.573935e-109 2.594864e-109 6.487300e-109 2.015087e-109 2.316533e-109
#>  [661] 8.294012e-109 2.013318e-109 2.575372e-108 2.826551e-109 3.023179e-109
#>  [666] 2.264734e-109 4.386929e-109 2.332277e-109 2.533725e-109 2.879985e-109
#>  [671] 1.024487e-108 1.963574e-109 3.416378e-109 2.056089e-109 2.106935e-109
#>  [676] 2.324030e-109 1.982725e-109 2.799137e-109 1.973301e-109 7.139375e-109
#>  [681] 2.335779e-109 7.150534e-109 4.616950e-109 2.435154e-109 3.694778e-109
#>  [686] 1.413466e-107 3.018572e-109 1.961579e-109 1.339648e-107 3.175320e-109
#>  [691] 1.597506e-108 4.804366e-109 2.006831e-109 2.800127e-109 2.134403e-109
#>  [696] 2.486693e-109 1.960798e-109 1.990640e-109 4.255113e-109 2.569232e-109
#>  [701] 1.067035e-108 9.842653e-109 3.506547e-109 2.663094e-109 2.997204e-109
#>  [706] 8.680404e-109 8.284838e-109 2.519930e-109 4.398421e-109 1.646823e-108
#>  [711] 2.352367e-108 3.363270e-108 2.470481e-109 5.007822e-109 3.333455e-109
#>  [716] 2.082573e-108 5.203295e-109 3.099140e-109 2.481217e-109 7.278084e-109
#>  [721] 2.213048e-109 2.837597e-109 2.313389e-109 2.302361e-108 3.641218e-109
#>  [726] 2.024976e-109 2.725640e-108 1.030469e-107 1.960826e-109 3.222560e-109
#>  [731] 3.373745e-109 1.830334e-107 3.059251e-109 6.804835e-109 7.015340e-109
#>  [736] 2.416063e-109 4.990935e-109 7.283929e-109 4.287563e-109 3.011348e-109
#>  [741] 6.696124e-109 2.194233e-109 2.123759e-109 1.961932e-109 1.018307e-108
#>  [746] 1.527963e-108 2.686192e-109 5.496890e-109 1.230256e-108 2.443630e-109
#>  [751] 4.584018e-109 6.778476e-109 2.226289e-109 2.237416e-109 2.044610e-109
#>  [756] 5.162432e-109 1.641329e-108 5.358254e-108 1.964851e-109 1.289315e-108
#>  [761] 2.006355e-109 2.801213e-109 2.396878e-108 3.736734e-109 1.115602e-108
#>  [766] 4.926197e-108 3.159234e-109 6.097461e-109 3.008671e-109 2.192901e-109
#>  [771] 9.226303e-109 4.565150e-109 3.264020e-109 1.240541e-107 2.043169e-109
#>  [776] 1.960903e-109 3.198356e-109 4.235337e-109 4.529154e-109 3.835080e-109
#>  [781] 2.330573e-109 2.812726e-109 4.394364e-109 3.214757e-109 1.318845e-108
#>  [786] 2.069638e-109 3.907214e-109 2.207616e-109 9.444761e-109 1.973898e-109
#>  [791] 4.067852e-109 5.165596e-109 6.286928e-108 1.967734e-109 8.994980e-109
#>  [796] 2.139335e-109 7.479541e-109 1.630570e-108 2.669953e-108 5.546984e-109
#>  [801] 2.196259e-109 2.786633e-108 2.818076e-109 3.704276e-109 2.079102e-109
#>  [806] 2.525799e-109 2.431229e-109 2.134244e-109 8.841216e-109 1.844474e-108
#>  [811] 2.744635e-109 9.487620e-109 6.535119e-109 3.210713e-109 4.588724e-109
#>  [816] 2.029494e-109 3.713084e-109 4.339590e-108 2.131728e-109 1.657312e-108
#>  [821] 7.634208e-109 8.757432e-109 2.086831e-109 5.302868e-109 5.053578e-109
#>  [826] 4.865613e-109 1.990658e-109 2.189755e-109 1.975904e-109 2.226910e-109
#>  [831] 6.119647e-109 6.827922e-109 1.542649e-108 3.338218e-109 1.988182e-109
#>  [836] 2.086505e-109 2.570074e-109 2.658195e-109 6.385402e-109 3.258957e-109
#>  [841] 4.053824e-109 3.936031e-109 3.829382e-109 3.199652e-109 6.442196e-109
#>  [846] 3.499266e-109 2.025642e-109 3.931583e-109 3.950668e-109 1.982554e-109
#>  [851] 2.112212e-109 1.662691e-108 1.976540e-109 3.933852e-109 2.185944e-109
#>  [856] 1.411296e-108 1.991673e-109 2.301379e-109 3.948415e-109 7.930993e-109
#>  [861] 4.661036e-109 2.007645e-109 1.636239e-108 2.010017e-109 3.683944e-109
#>  [866] 5.848788e-109 9.797956e-109 2.200874e-109 3.402363e-109 6.143183e-109
#>  [871] 2.072637e-109 4.251982e-109 9.038933e-109 3.827022e-109 3.017425e-109
#>  [876] 2.425560e-109 7.780798e-109 2.921925e-108 2.046561e-109 1.991406e-109
#>  [881] 5.894034e-109 3.362672e-109 9.397057e-109 2.180230e-109 5.268742e-109
#>  [886] 1.982950e-109 4.062095e-109 1.985416e-109 2.117899e-109 2.453407e-109
#>  [891] 1.193556e-108 2.396822e-109 3.128772e-109 2.006690e-109 1.974107e-109
#>  [896] 3.581975e-109 2.346778e-108 2.362153e-108 5.088843e-109 2.465629e-109
#>  [901] 3.193345e-108 2.942791e-109 9.860767e-109 5.940812e-109 1.011434e-108
#>  [906] 2.602846e-108 4.695103e-109 6.488391e-109 2.757604e-109 3.024874e-109
#>  [911] 2.155945e-109 3.697598e-109 3.315497e-109 8.642036e-109 3.198261e-107
#>  [916] 1.961297e-109 2.130550e-109 1.971795e-109 2.630524e-109 3.566047e-109
#>  [921] 3.741438e-109 3.142361e-109 7.676393e-109 2.065630e-109 2.088154e-109
#>  [926] 1.989146e-109 1.971476e-109 1.990288e-109 2.128100e-109 1.662746e-108
#>  [931] 3.925536e-109 5.358402e-109 6.626341e-109 1.855772e-107 3.216673e-109
#>  [936] 1.972169e-109 3.549057e-109 3.381289e-109 4.243518e-109 6.189428e-109
#>  [941] 2.083230e-109 6.328761e-109 2.004547e-109 2.990343e-109 5.119696e-108
#>  [946] 5.276464e-109 1.813646e-108 2.381800e-109 1.962169e-109 4.166973e-109
#>  [951] 2.191947e-109 5.535772e-109 2.463681e-109 2.857233e-109 2.667456e-109
#>  [956] 2.021746e-109 2.520095e-109 1.961378e-109 1.795608e-108 5.854733e-109
#>  [961] 1.143401e-108 3.814583e-109 3.831694e-109 6.422200e-109 1.988334e-108
#>  [966] 1.980416e-109 2.996192e-107 1.986438e-109 2.059446e-109 2.104048e-109
#>  [971] 2.184950e-109 2.126620e-109 2.014942e-109 1.840820e-108 7.499020e-109
#>  [976] 3.384696e-109 1.005339e-107 1.972082e-109 2.845878e-109 1.193667e-108
#>  [981] 2.475915e-109 7.680753e-109 2.153140e-109 1.114040e-108 2.355389e-109
#>  [986] 1.995344e-109 2.789190e-109 2.303000e-109 3.550402e-109 7.142855e-109
#>  [991] 2.861791e-109 3.452192e-109 5.975449e-108 1.166951e-108 1.966777e-109
#>  [996] 1.040156e-107 2.365352e-109 3.957772e-109 2.902790e-109 3.467428e-108
susie_get_niter(s)
#> [1] 4
susie_get_pip(s)
#>    [1] 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>   [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>   [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [112] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [149] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [223] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [260] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [297] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [334] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [371] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [408] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [445] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [482] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [519] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [556] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [593] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [630] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [667] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [704] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [741] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [778] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [815] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [852] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [889] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [926] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [963] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [1000] 0
susie_get_lfsr(s)
#>  [1] 0.0 0.0 0.0 0.0 0.2 0.2 0.2 0.2 0.2 0.2