Sum of Single Effects (SuSiE) Regression using summary statistics

susie_rss performs variable selection under a sparse Bayesian multiple linear regression of \(Y\) on \(X\) using only the z-scores from standard univariate regression of \(Y\) on each column of \(X\), and an estimate \(R\) of the correlation matrix between columns of \(X\). It does this by combining the "RSS likelihood" from Zhu and Stephens (2017) with the Sum of Single Effects" model from Wang et al (2020).

susie_rss(
  z,
  R,
  z_ld_weight = 0,
  L = 10,
  prior_variance = 50,
  estimate_prior_variance = TRUE,
  ...
)

Arguments

z
A p-vector of z scores.

R
A p by p symmetric, positive semidefinite correlation matrix.

z_ld_weight
This parameter is included for backwards compatibility with previous versions of the function, but it is no longer recommended to use a non-zero value. If z_ld_weight > 0, the matrix R used in the model is adjusted to be cov2cor((1-w)*R + w*tcrossprod(z)), where w = z_ld_weight.

L
Maximum number of components (nonzero coefficients) in the susie regression model. If L is larger than the number of covariates, p, L is set to p.

prior_variance
The prior variance(s) for the non-zero element of \(b_l\). It is either a scalar, or a vector of length L. When estimate_prior_variance = TRUE (highly recommended) this simply provides an initial value for the prior variance, and the default value of 50 is simply intended to be a large initial value.

estimate_prior_variance
If estimate_prior_variance = TRUE, which is highly recommended, the prior variance is estimated (this is a separate parameter for each of the L effects). If provided, prior_variance is then used as an initial value for the optimization. When estimate_prior_variance = FALSE (not recommended) the prior variance for each of the L effects is determined by the value supplied to prior_variance.

...
Other parameters to be passed to susie_suff_stat.

Value

A "susie" object with the following elements:

alpha

An L by p matrix of posterior inclusion probabilites.

mu

An L by p matrix of posterior means, conditional on inclusion.

mu2

An L by p matrix of posterior second moments, conditional on inclusion.

lbf

log-Bayes Factor for each single effect.

lbf_variable

log-Bayes Factor for each variable and single effect.

V

Prior variance of the non-zero elements of b.

elbo

The value of the variational lower bound, or “ELBO” (objective function to be maximized), achieved at each iteration of the IBSS fitting procedure.

Rr

A vector of length p, equal to R %*% colSums(alpha*mu).

sets

Credible sets estimated from model fit; see susie_get_cs for details.

pip

A vector of length p giving the (marginal) posterior inclusion probabilities for all p covariates.

niter

Number of IBSS iterations that were performed.

converged

TRUE or FALSE indicating whether the IBSS converged to a solution within the chosen tolerance level.

Details

In some applications, particularly genetic applications, it is desired to fit a regression model (\(Y = X\tilde{b} + E\) say, which we refer to as "the original regression model" or ORM) without access to the actual values of \(Y\) and \(X\), but given only some summary statistics. susie_rss assumes the availability of \(z\) scores from standard univariate regression of \(Y\) on each column of \(X\), and an estimate \(R\) of the correlation matrix between columns of \(X\) (\(R\) is sometimes called the LD matrix in genetic applications). See Zhu and Stephens (2017), and references therein, for further background.

The susie_rss function is based on the model (2.10) from Zhu and Stephens, \(z | R, b ~ N(Rb,R)\) where \(b\) is a vector of length p representing the effects to be estimated. The effect \(b_j\) is simply a multiple of the coefficient \(\tilde{b}_j\) in the ORM, and so \(b_j\) is non-zero if and only if \(\tilde{b}_j\) is non-zero. In this sense the variable selection problem in this model is the same as the variable selection problem in the ORM, and so the credible sets and PIPs computed by susie_rss can be interpreted as credible sets and PIPs for the ORM. However, converting posterior estimates of \(b_j\) to estimates of \(\tilde{b}_j\) would require computation of the scaling factor (not done here).

More precisely, susie_rss assumes the log-likelihood for \(b\) is \(l(b; z,R) = -0.5(b'Rb - 2z'b)\), which is equivalent to model (2.10) from Zhu and Stephens if \(R\) is invertible, but does not require \(R\) to be invertible. It combines this likelihood with the “susie prior” which assumes that \(b = \sum_{l=1}^L b_l\) where each \(b_l\) is a vector of length p with exactly one non-zero element; see susie and Wang et al (2020) for details.

In practice, this is accomplished by calling susie_suff_stat with XtX = R and Xty = z, and fixing residual_variance = 1. (Values for n and yty are also required by susie_suff_stat. They do not affect inference when the residual variance is fixed, but they do affect the interpretation of scaled_prior_variance; we set n=2, yty=1 so that \(var(y) = yty/(n-1) = 1\).) Additional arguments to be passed to susie_suff_stat can be provided via ....

References

G. Wang, A. Sarkar, P. Carbonetto and M. Stephens (2020). A simple new approach to variable selection in regression, with application to genetic fine-mapping. Journal of the Royal Statistical Society, Series B 82 doi: 10.1101/501114 .

X. Zhu and M. Stephens (2017). Bayesian large-scale multiple regression with summary statistics from genome-wide association studies. Annals of Applied Statistics 11, 1561-1592.

Examples

set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1:4] = 1
X = matrix(rnorm(n*p),nrow = n,ncol = p)
X = scale(X,center = TRUE,scale = TRUE)
y = drop(X %*% beta + rnorm(n))

input_ss = compute_suff_stat(X,y,standardize = TRUE)
ss   = univariate_regression(X,y)
R    = with(input_ss,cov2cor(XtX))
zhat = with(ss,betahat/sebetahat)
res  = susie_rss(zhat,R,L = 10)